% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_prediction_cmprsk.R
\name{as_prediction_cmprsk}
\alias{as_prediction_cmprsk}
\alias{as_prediction_cmprsk.PredictionCompRisks}
\alias{as_prediction_cmprsk.data.frame}
\title{Convert to a Competing Risk Prediction}
\usage{
as_prediction_cmprsk(x, ...)

\method{as_prediction_cmprsk}{PredictionCompRisks}(x, ...)

\method{as_prediction_cmprsk}{data.frame}(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
\link{PredictionCompRisks}.
}
\description{
Convert object to a \link{PredictionCompRisks}.
}
\examples{
library(mlr3)
task = tsk("pbc")
learner = lrn("cmprsk.aalen")
learner$train(task)
p = learner$predict(task)

# convert to a data.table
tab = as.data.table(p)

# convert back to a Prediction
as_prediction_cmprsk(tab)
}
