% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_dens.R
\name{as_task_dens}
\alias{as_task_dens}
\alias{as_task_dens.TaskDens}
\alias{as_task_dens.data.frame}
\alias{as_task_dens.DataBackend}
\title{Convert to a Density Task}
\usage{
as_task_dens(x, ...)

\method{as_task_dens}{TaskDens}(x, clone = FALSE, ...)

\method{as_task_dens}{data.frame}(x, id = deparse(substitute(x)), ...)

\method{as_task_dens}{DataBackend}(x, id = deparse(substitute(x)), ...)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert, e.g. a \code{data.frame()}.}

\item{...}{(\code{any})\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}

\item{id}{(\code{character(1)})\cr
Id for the new task.
Defaults to the (deparsed and substituted) name of \code{x}.}
}
\description{
Convert object to a density task (\link{TaskDens}).
}
