% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_surv}
\alias{assert_surv}
\title{Assert survival object}
\usage{
assert_surv(
  x,
  len = NULL,
  any.missing = TRUE,
  null.ok = FALSE,
  .var.name = vname(x)
)
}
\arguments{
\item{x}{Object to check}

\item{len}{If non-NULL checks object is length \code{len}}

\item{any.missing}{If \code{FALSE} then errors if there are any NAs in \code{x}}

\item{null.ok}{If \code{FALSE} then errors if \code{x} is NULL, otherwise passes}

\item{.var.name}{Optional variable name to return if assertion fails}
}
\description{
Asserts \code{x} is a \link[survival:Surv]{survival::Surv} object with added checks
}
