% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.LearnerSurv}
\alias{autoplot.LearnerSurv}
\title{Visualization of fitted \code{LearnerSurv} objects}
\usage{
\method{autoplot}{LearnerSurv}(object, task, fun = "survival", row_ids = NULL, newdata, ...)
}
\description{
Wrapper around \code{predict.LearnerSurv} and `plot.
}
\examples{
library(mlr3)
task = tsk("rats")

# Prediction Error Curves for prediction object
learn = lrn("surv.coxph")
autoplot(learn)

autoplot(object, task, "survival", ind = 1:10)
autoplot(object, task, "survival", newdata = ndf)


}
