% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.PredictionSurv}
\alias{autoplot.PredictionSurv}
\title{Plot for PredictionSurv}
\usage{
\method{autoplot}{PredictionSurv}(
  object,
  type = "calib",
  times = NULL,
  row_ids = NULL,
  cuts = 11L,
  theme = theme_minimal(),
  ...
)
}
\arguments{
\item{object}{(\link{PredictionSurv}).}

\item{type}{(\code{character(1)}) \cr
Type of the plot, see Description.}

\item{times}{(\code{numeric()}) \cr
If \code{type = "calib"} then \code{times} is the values on the x-axis to plot over.
if \code{NULL} uses all time points from the predicted survival matrix (\code{object$data$distr}).}

\item{row_ids}{(\code{integer()}) \cr
If \code{type = "isd"}, specific observation ids (from the test set) for which
we draw their predicted survival distributions.}

\item{cuts}{(\code{integer(1)}) \cr
Number of cuts in \eqn{(0,1)} to plot \code{dcalib} over, default is \code{11}.}

\item{theme}{(\code{\link[ggplot2:theme]{ggplot2::theme()}})\cr
The \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}} is applied by default to all plots.}

\item{...}{(\code{any}):
Additional arguments, currently unused.}
}
\description{
Generates plots for \link{PredictionSurv}, depending on argument \code{type}:
\itemize{
\item \code{"calib"} (default): \strong{Calibration plot} comparing the average predicted
survival distribution (\code{Pred}) to a Kaplan-Meier prediction (\code{KM}), this is
\emph{not} a comparison of a stratified \code{crank} or \code{lp}.
\item \code{"dcalib"}: \strong{Distribution calibration plot}.
A model is considered D-calibrated if, for any given quantile \code{p}, the
proportion of observed outcomes occurring before the predicted time quantile,
matches \code{p}. For example, 50\% of events should occur before the predicted
median survival time (i.e. the time corresponding to a predicted survival
probability of 0.5).
This means that the resulting line plot will lie close to the straight line
y = x.
Note that we impute \code{NA}s from the predicted quantile function with the
maximum observed outcome time.
\item \code{"isd"}: Plot the predicted \strong{i}ndividual \strong{s}urvival \strong{d}istributions
(survival curves) for observations from the test set.
}
}
\section{Notes}{

\enumerate{
\item \code{object} must have a \code{distr} prediction, as all plot \code{type}s use the
predicted survival distribution/matrix.
\item \code{type = "dcalib"} is drawn a bit differently from Haider et al. (2020),
though its still conceptually the same.
}
}

\examples{
\dontshow{if (mlr3misc::require_namespaces(c("mlr3viz", "ggplot2"), quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mlr3)
library(mlr3proba)
library(mlr3viz)

learner = lrn("surv.coxph")
task = tsk("gbcs")
p = learner$train(task, row_ids = 1:300)$predict(task, row_ids = 301:400)

# calibration by comparison of average prediction to Kaplan-Meier
autoplot(p)

# same as above, use specific time points
autoplot(p, times = seq(1, 1000, 5))

# Distribution-calibration (D-Calibration)
autoplot(p, type = "dcalib")

# Predicted survival curves (all observations)
autoplot(p, type = "isd")

# Predicted survival curves (specific observations)
autoplot(p, type = "isd", row_ids = c(301, 351, 399))
\dontshow{\}) # examplesIf}
}
\references{
Haider, Humza, Hoehn, Bret, Davis, Sarah, Greiner, Russell (2020).
\dQuote{Effective Ways to Build and Evaluate Individual Survival Distributions.}
\emph{Journal of Machine Learning Research}, \bold{21}(85), 1--63.
\url{https://jmlr.org/papers/v21/18-772.html}.
}
