% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crankcompositor.R
\name{crankcompositor}
\alias{crankcompositor}
\title{Compose a Crank Predict Type for Survival Learners}
\usage{
crankcompositor(learner, method = "mean", param_vals = list())
}
\arguments{
\item{learner}{\link{LearnerSurv} object for which a \code{crank} is composed (or over-written)}

\item{method}{One of \code{mean} or \code{median}. Used to determine how \code{crank} is estimated from the
predicted \code{distr}.}

\item{param_vals}{Additional parameters to pass to the \code{learner}.}
}
\value{
\link{Graph}
}
\description{
This is a wrapper around the \link{mlr_pipeops_crankcompose} pipe operation, which
simplifies graph creation.
}
\details{
For full details see \link{PipeOpCrankCompositor}.
}
\examples{
library("mlr3")
library("mlr3pipelines")

ranger.crank = crankcompositor(learner = lrn("surv.ranger"),
                            method = "median")
resample(tsk("rats"), ranger.crank, rsmp("cv", folds = 2))$predictions()
}
