% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrcompositor.R
\name{distrcompositor}
\alias{distrcompositor}
\title{Compose a Distr Predict Type for Survival Learners}
\usage{
distrcompositor(
  learner,
  estimator = "kaplan",
  form = "aft",
  param_vals = list()
)
}
\arguments{
\item{learner}{\link{LearnerSurv} object for which a \code{distr} is composed (or over-written)}

\item{estimator}{One of \code{kaplan} or \code{nelson}, corresponding to the Kaplan-Meier and Nelson-Aalen
estimators respectively. Used to estimate the baseline survival distribution. Partial matching allowed.}

\item{form}{One of \code{aft}, \code{ph}, or \code{po}, corresponding to accelerated failure time, proportional hazards,
and proportional odds respectively. Used to determine the form of the composed survival distribution.}

\item{param_vals}{Additional parameters to pass to the \code{learner}.}
}
\value{
\link{Graph}
}
\description{
This is a wrapper around the \link{mlr_pipeops_distrcompose} pipe operation, which
simplifies graph creation.
}
\details{
For full details see \link{mlr_pipeops_distrcompose}.
}
\examples{
library("mlr3")
library("mlr3pipelines")

cox.distr = distrcompositor(learner = lrn("surv.coxph"),
                            estimator = "kaplan",
                            form = "aft")
resample(tsk("rats"), cox.distr, rsmp("cv", folds = 2))
}
