% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breslow.R
\name{.cbhaz_breslow}
\alias{.cbhaz_breslow}
\title{Breslow's Cumulative Baseline Hazard}
\usage{
.cbhaz_breslow(times, status, lp, eval_times)
}
\arguments{
\item{times}{(\code{numeric()})\cr Vector of times (train set).}

\item{status}{(\code{numeric()})\cr Vector of status indicators (train set).
For each observation in the train set, this should be 0 (alive/censored) or
1 (dead).}

\item{lp}{(\code{numeric()})\cr Vector of linear predictors (train set).
These are the relative score predictions from a proportional hazards model
on the train set.}

\item{eval_times}{(\code{numeric()})\cr Vector of times to compute the cumulative
baseline hazard estimations. If \code{NULL} (default), the unique and sorted
\code{times} from the train set will be used, otherwise the unique and sorted
\code{eval_times}.}
}
\value{
A vector of cumulative baseline hazards, one per (increasing) time
point. The times are added as names in the output vector.
}
\description{
Computes the cumulative baseline hazard using Breslow's (1972) estimator.
Assumes risk predictions are derived from a \strong{proportional hazards}
regression model.
}
\details{
Breslow's approach uses a non-parametric maximum likelihood estimation of the
cumulative baseline hazard function:

\deqn{\hat{H}_0(t) = \sum_{i=1}^n{\frac{I(T_i \le t)\delta_i}
{\sum\nolimits_{j \in R_i}e^{lp_j}}}}

where:
\itemize{
\item \eqn{\hat{H}_0(t)} is the cumulative baseline hazard
\item \eqn{t} is the vector of time points
\item \eqn{n} is number of events
\item \eqn{T} is the vector of event times
\item \eqn{\delta} is the status indicator (event or censoring)
\item \eqn{R_i} is the risk set (number of individuals at risk just before
event \eqn{i})
\item \eqn{lp_j} is the risk prediction (linear predictor) of individual \eqn{j}
(who is part of the risk set \eqn{R_i}).
}

For similar implementations, see \code{gbm::basehaz.gbm()}, \code{C060::basesurv()} and
\code{xgboost.surv::sgb_bhaz()}.
}
\examples{
task = tsk("rats")
part = partition(task, ratio = 0.8)

learner = lrn("surv.coxph")
learner$train(task, part$train)
p_train = learner$predict(task, part$train)

base_haz = .cbhaz_breslow(times = task$times(part$train),
  status = task$status(part$train), lp = p_train$lp)
head(base_haz)

}
\references{
Cox DR (1972).
\dQuote{Regression Models and Life-Tables.}
\emph{Journal of the Royal Statistical Society: Series B (Methodological)}, \bold{34}(2), 187--202.
\doi{10.1111/j.2517-6161.1972.tb00899.x}.

Lin, Y. D (2007).
\dQuote{On the Breslow estimator.}
\emph{Lifetime Data Analysis}, \bold{13}(4), 471-480.
\doi{10.1007/S10985-007-9048-Y/METRICS}.
}
