% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grace}
\alias{grace}
\title{GRACE 1000 Dataset}
\format{
\describe{
\item{id}{Identification Code}
\item{days}{Follow up time.}
\item{death}{Censoring indicator. 1 = Death. 0 = Censored.}
\item{revasc}{Revascularization Performed. 1 = Yes. 0 = No.}
\item{revascdays}{Days to revascularization after admission.}
\item{los}{Length of hospital stay (days).}
\item{age}{Age at admission (years).}
\item{sysbp}{Systolic blood pressure on admission (mm Hg).}
\item{stchange}{ST-segment deviation on index ECG. 1 = Yes. 0 = No.}
}
}
\source{
\url{ftp://ftp.wiley.com/public/sci_tech_med/survival}
}
\usage{
grace
}
\description{
grace dataset from Hosmer et al.
}
\references{
Hosmer, D.W. and Lemeshow, S. and May, S. (2008)
Applied Survival Analysis: Regression Modeling of Time to Event Data: Second
Edition, John Wiley and Sons Inc., New York, NY
}
\keyword{datasets}
