% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDens.R
\name{LearnerDens}
\alias{LearnerDens}
\title{Density Learner}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Learner}.}
\description{
This Learner specializes \link{Learner} for density estimation problems.
}
\section{Construction}{
\preformatted{l = LearnerDens$new(id, param_set = ParamSet$new(), predict_types = "cdf",
feature_types = character(), properties = character(), data_formats = "data.table",
packages = character())
}

For a description of the arguments, see \link{Learner}.
\code{task_type} is set to \code{"dens"}.

Possible values for \code{predict_types} are passed to and converted by \link{PredictionDens}:
\itemize{
\item \code{"pdf"}: Evaluates the estimated probability density function for each value in the test set.
\item \code{"cdf"}: Evaluates the estimated cumulative distribution function for each value in the test set.
}
}

\section{Fields}{

See \link{Learner}.
}

\section{Methods}{

See \link{Learner}.
}

\examples{
library(mlr3)
# get all density learners from mlr_learners:
lrns = mlr_learners$mget(mlr_learners$keys("^dens"))
names(lrns)

# get a specific learner from mlr_learners:
mlr_learners$get("dens.hist")
lrn("dens.hist")
}
\seealso{
Other Learner: 
\code{\link{LearnerSurv}}
}
\concept{Learner}
