% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensHistogram.R
\name{LearnerDensHistogram}
\alias{LearnerDensHistogram}
\alias{mlr_learners_dens.hist}
\title{Histogram Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[graphics:hist]{graphics::hist()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensHistogram$new()
mlr_learners$get("dens.hist")
lrn("dens.hist")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \verb{pdf, cdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{graphics} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLogLocfit}}
}
\concept{density estimators}
