% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDE.R
\name{LearnerDensKDE}
\alias{LearnerDensKDE}
\alias{mlr_learners_dens.kde}
\title{Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls kernels implemented in \CRANpkg{distr6} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensKDE$new()
mlr_learners$get("dens.kde")
lrn("dens.kde")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
