% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDEgk.R
\name{LearnerDensKDEgk}
\alias{LearnerDensKDEgk}
\alias{mlr_learners_dens.kdeGK}
\title{Gaussian Univariate Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[GenKern:KernSec]{GenKern::KernSec()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensKDEgk$new()
mlr_learners$get("dens.kdeGK")
lrn("dens.kdeGK")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{GenKern} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDE}}
}
\concept{density estimators}
