% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDEkd.R
\name{LearnerDensKDEkd}
\alias{LearnerDensKDEkd}
\alias{mlr_learners_dens.kdeKD}
\title{Kerdiest Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[kerdiest:kde]{kerdiest::kde()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensKDEkd$new()
mlr_learners$get("dens.kdeKD")
lrn("dens.kdeKD")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{kerdiest} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
