% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDEks.R
\name{LearnerDensKDEks}
\alias{LearnerDensKDEks}
\alias{mlr_learners_dens.kdeKS}
\title{KS Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[ks:kde]{ks::kde()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensKDEks$new()
mlr_learners$get("dens.kdeKS")
lrn("dens.kdeKS")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Properties: \code{-}
\item Packages: \CRANpkg{ks} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensLogspline}},
\code{\link{LearnerDensMixed}},
\code{\link{LearnerDensNonparametric}},
\code{\link{LearnerDensPenalized}},
\code{\link{LearnerDensPlugin}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
