% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDEnp.R
\name{LearnerDensKDEnp}
\alias{LearnerDensKDEnp}
\alias{mlr_learners_dens.kdeNP}
\title{Mixed Data Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[np:npudens]{np::npudens()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensKDEnp$new()
mlr_learners$get("dens.kdeNP")
lrn("dens.kdeNP")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{np} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDE}}
}
\concept{density estimators}
