% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDEsm.R
\name{LearnerDensKDEsm}
\alias{LearnerDensKDEsm}
\alias{mlr_learners_dens.kdeSM}
\title{Nonparametric Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[sm:sm.density]{sm::sm.density()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensKDEsm$new()
mlr_learners$get("dens.kdeSM")
lrn("dens.kdeSM")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{sm} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDE}}
}
\concept{density estimators}
