% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensLogLocfit.R
\name{LearnerDensLogLocfit}
\alias{LearnerDensLogLocfit}
\alias{mlr_learners_dens.logLF}
\title{Local Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[locfit:density.lf]{locfit::density.lf()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensLogLocfit$new()
mlr_learners$get("dens.logLF")
lrn("dens.logLF")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{locfit} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDE}}
}
\concept{density estimators}
