% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensLogspline.R
\name{LearnerDensLogspline}
\alias{LearnerDensLogspline}
\alias{mlr_learners_dens.logspline}
\title{Logspline Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[logspline:logspline]{logspline::logspline()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensLogspline$new()
mlr_learners$get("dens.logspline")
lrn("dens.logspline")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \verb{pdf, cdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{logspline} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensMixed}},
\code{\link{LearnerDensPenalized}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
