% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensMixed.R
\name{LearnerDensMixed}
\alias{LearnerDensMixed}
\alias{mlr_learners_dens.mixed}
\title{Mixed Data Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[np:npudens]{np::npudens()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensMixed$new()
mlr_learners$get("dens.mixed")
lrn("dens.mixed")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{np} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensLogspline}},
\code{\link{LearnerDensNonparametric}},
\code{\link{LearnerDensPenalized}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
