% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensPenGss.R
\name{LearnerDensPenGss}
\alias{LearnerDensPenGss}
\alias{mlr_learners_dens.penGSS}
\title{Smoothing Splines Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[gss:ssden]{gss::ssden()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensPenGss$new()
mlr_learners$get("dens.penGSS")
lrn("dens.penGSS")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{gss} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDE}}
}
\concept{density estimators}
