% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensPenLP.R
\name{LearnerDensPenLP}
\alias{LearnerDensPenLP}
\alias{mlr_learners_dens.penLP}
\title{Logspline Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[logspline:logspline]{logspline::logspline()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensPenLP$new()
mlr_learners$get("dens.penLP")
lrn("dens.penLP")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{logspline} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDE}}
}
\concept{density estimators}
