% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensPenPD.R
\name{LearnerDensPenPD}
\alias{LearnerDensPenPD}
\alias{mlr_learners_dens.penPD}
\title{Penalized Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[pendensity:pendensity]{pendensity::pendensity()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensPenPD$new()
mlr_learners$get("dens.penPD")
lrn("dens.penPD")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \verb{pdf, cdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{pendensity} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
