% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensPlugin.R
\name{LearnerDensPlugin}
\alias{LearnerDensPlugin}
\alias{mlr_learners_dens.plug}
\title{Plug-In Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[plugdensity:plugin.density]{plugdensity::plugin.density()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensPlugin$new()
mlr_learners$get("dens.plug")
lrn("dens.plug")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{plugdensity} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDE}}
}
\concept{density estimators}
