% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensSpline.R
\name{LearnerDensSpline}
\alias{LearnerDensSpline}
\alias{mlr_learners_dens.spline}
\title{Smoothing Splines Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \code{\link[gss:ssden]{gss::ssden()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensSpline$new()
mlr_learners$get("dens.spline")
lrn("dens.spline")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \verb{pdf, cdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{gss} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensLogspline}},
\code{\link{LearnerDensPenalized}}
}
\concept{density estimators}
