% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvBlackboost.R
\name{LearnerSurvBlackboost}
\alias{LearnerSurvBlackboost}
\alias{mlr_learners_surv.blackboost}
\title{Gradient Boosting with Regression Trees Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Gradient boosting for optimizing arbitrary loss functions where regression trees are
utilized as base-learners.
Calls \code{\link[mboost:blackboost]{mboost::blackboost()}} from package \CRANpkg{mboost}.
}
\details{
The \code{distr} return type is composed by using\code{\link[mboost:survFit]{mboost::survFit()}} which assumes a PH fit with
a Breslow estimator.
The \code{crank} return type is defined by the expectation of the survival distribution. \cr
The \code{lp} return type is given by \code{\link[mboost:predict.blackboost]{mboost::predict.blackboost()}}.

If the value given to the \code{Family} parameter is "custom.family" then an object of class
\code{\link[mboost:Family]{mboost::Family()}} needs to be passed to the \code{custom.family} parameter.
}
\section{Construction}{
\preformatted{LearnerSurvBlackboost$new()
mlr_learners$get("surv.blackboost")
lrn("surv.blackboost")
}
}

\examples{
library(mlr3)
task = tgen("simsurv")$generate(200)
learner = lrn("surv.blackboost")
resampling = rsmp("cv", folds = 3)
resample(task, learner, resampling)
}
\references{
Peter Buehlmann and Torsten Hothorn (2007),
Boosting algorithms: regularization, prediction and model fitting.
Statistical Science, 22(4), 477–505.

Torsten Hothorn, Kurt Hornik and Achim Zeileis (2006).
Unbiased recursive partitioning: A conditional inference framework.
Journal of Computational and Graphical Statistics, 15(3), 651–674.

Yoav Freund and Robert E. Schapire (1996), Experiments with a new boosting algorithm.
In Machine Learning: Proc. Thirteenth International Conference, 148–156.

Jerome H. Friedman (2001), Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29, 1189–1232.

Greg Ridgeway (1999), The state of boosting.
Computing Science and Statistics, 31, 172–181.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
