% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvCVGlmnet.R
\name{LearnerSurvCVGlmnet}
\alias{LearnerSurvCVGlmnet}
\alias{mlr_learners_surv.cvglmnet}
\title{Cross-Validated GLM with Elastic Net Regularization Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Generalized linear models with elastic net regularization and k-fold cross-validation for lambda.
Calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} from package \CRANpkg{glmnet}.
}
\details{
The \code{distr} return type is composed by using the Cox formula \eqn{S(t) = S0(t)^exp(lp)} where
\eqn{S0} is the baseline survival function and \eqn{lp} is the predicted linear predictor. The
baseline survival is estimated with either Kaplan-Meier or Nelson-Aalen, depending on the
\code{estimator} hyper-parameter, using \code{\link[survival:survfit]{survival::survfit()}}.\cr
The \code{crank} return type is defined by the expectation of the survival distribution. \cr
The \code{lp} return type is given natively by \code{\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet()}}.

See \link{LearnerSurvCVGlmnet} for the glmnet implemented in \code{\link[glmnet:glmnet]{glmnet::glmnet()}} without internal cross-validation.
Tuning using the internal optimizer in \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} may be more efficient when tuning
lambda only, however for tuning multiple hyperparameters, \CRANpkg{mlr3tuning} and \code{\link[glmnet:glmnet]{glmnet::glmnet()}} will
likely give better results.
}
\section{Construction}{
\preformatted{LearnerSurvCVGlmnet$new()
mlr_learners$get("surv.cvglmnet")
lrn("surv.cvglmnet")
}
}

\examples{
library(mlr3)
task = tgen("simsurv")$generate(200)
learner = lrn("surv.cvglmnet")
resampling = rsmp("cv", folds = 3)
resample(task, learner, resampling)
}
\references{
Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010).
Regularization Paths for Generalized Linear Models via Coordinate Descent.
Journal of Statistical Software, 33(1), 1-22.
\doi{10.18637/jss.v033.i01}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
