% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvFlexible.R
\name{LearnerSurvFlexible}
\alias{LearnerSurvFlexible}
\alias{mlr_learners_surv.flexible}
\title{Flexible Parametric Spline Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Calls \code{\link[survival:survreg]{survival::survreg()}}.
\itemize{
\item lp is predicted by using an internally defined \code{predict} method, see details.
\item distr is predicted by using an internally defined \code{predict} method, see details.
\item \code{crank} is identical to \code{lp}.
}
}
\details{
The predict method is based on \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}} but is quicker and more
efficient for adapting the return type to be compatible with \CRANpkg{distr6}. A large amount of the
predict method is copied from \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}} and credit goes to Christopher Jackson
for creating/maintaining \CRANpkg{flexsurv}.

\code{lp} is predicted using the formula \eqn{lp = X\beta} where \eqn{X} are the variables in the test data set
and \eqn{\beta} are the fitted coefficients.

\code{distr} is predicted using the formula derived in \code{\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline()}} and combined with parameters
defined in \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}.
}
\section{Construction}{
\preformatted{LearnerSurvFlexible$new()
mlr_learners$get("surv.flexible")
lrn("surv.flexible")
}
}

\section{Meta Information}{

\itemize{
\item Type: "surv"
\item Predict Types: \verb{distr, lp, crank}
\item Feature Types: \verb{logical, integer, factor, numeric}
\item Packages: \CRANpkg{flexsurv} \CRANpkg{survival} \CRANpkg{distr6}
}
}

\references{
Royston, P. and Parmar, M. (2002).
Flexible parametric proportional-hazards and proportional-odds models for censored survival data,
with application to prognostic modelling and estimation of treatment effects.
Statistics in Medicine, 21(15), 2175-2197.
\doi{10.1002/sim.1203}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}

Other survival learners: 
\code{\link{LearnerSurvBlackboost}},
\code{\link{LearnerSurvCVGlmnet}},
\code{\link{LearnerSurvCoxPH}},
\code{\link{LearnerSurvGamboost}},
\code{\link{LearnerSurvGlmboost}},
\code{\link{LearnerSurvGlmnet}},
\code{\link{LearnerSurvKaplan}},
\code{\link{LearnerSurvMboost}},
\code{\link{LearnerSurvNelson}},
\code{\link{LearnerSurvParametric}},
\code{\link{LearnerSurvPenalized}},
\code{\link{LearnerSurvRandomForestSRC}},
\code{\link{LearnerSurvRanger}},
\code{\link{LearnerSurvRpart}},
\code{\link{LearnerSurvSVM}}
}
\concept{survival learners}
