% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvGbm.R
\name{LearnerSurvGbm}
\alias{LearnerSurvGbm}
\alias{mlr_learners_surv.gbm}
\title{Survival Gbm Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
A gradient boosting machine based \link{LearnerSurv} implemented in \code{\link[gbm:gbm]{gbm::gbm()}} in package \CRANpkg{gbm}.
Only Cox PH type models are supported for survival analysis.
}
\details{
The \code{lp} return type is given natively by predicting the linear predictor in \code{\link[gbm:predict.gbm]{gbm::predict.gbm()}}.\cr
}
\section{Construction}{
\preformatted{LearnerSurvGbm$new()
mlr_learners$get("surv.gbm")
lrn("surv.gbm")-
}
}

\references{
Brandon Greenwell, Bradley Boehmke, Jay Cunningham and GBM Developers (2019).
gbm: Generalized Boosted Regression Models. R package version 2.1.5.
\url{https://CRAN.R-project.org/package=gbm}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
