% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvKaplan.R
\name{LearnerSurvKaplan}
\alias{LearnerSurvKaplan}
\alias{mlr_learners_surv.kaplan}
\title{Kaplan-Meier Estimator Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Calls \code{\link[survival:survfit]{survival::survfit()}}.
\itemize{
\item distr is predicted by estimating the survival function with \code{\link[survival:survfit]{survival::survfit()}}
\item \code{crank} is predicted as the expectation of the survival distribution, \code{distr}
}
}
\section{Construction}{
\preformatted{LearnerSurvKaplan$new()
mlr_learners$get("surv.kaplan")
lrn("surv.kaplan")
}
}

\section{Meta Information}{

\itemize{
\item Type: "surv"
\item Predict Types: \verb{crank, distr}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{survival} \CRANpkg{distr6}
}
}

\references{
\cite{mlr3proba}{kaplan_1958}
}
\seealso{
Other survival learners: 
\code{\link{LearnerSurvGBM}},
\code{\link{LearnerSurvNelson}}
}
\concept{survival learners}
