% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvKaplanMeier.R
\docType{data}
\name{LearnerSurvKaplanMeier}
\alias{LearnerSurvKaplanMeier}
\alias{mlr_learners_surv.kaplan}
\title{Kaplan Meier Estimator Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Kaplan Meier estimator called from \code{\link[survival:survfit]{survival::survfit()}} in package \CRANpkg{survival}.
}
\details{
The \code{distr} return type is given natively by estimating the survival function with \code{\link[survival:survfit]{survival::survfit()}}.\cr
The \code{crank} return type is defined by the expectation of the survival distribution.
}
\section{Construction}{
\preformatted{LearnerSurvKaplanMeier$new()
mlr_learners$get("surv.kaplan")
lrn("surv.kaplan")
}
}

\references{
Kaplan, E. L. and Meier, P. (1958).
Nonparametric Estimation from Incomplete Observations.
Journal of the American Statistical Association, 53(282), 457-481.
\doi{10.2307/2281868}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
