% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvNelsonAalen.R
\docType{data}
\name{LearnerSurvNelsonAalen}
\alias{LearnerSurvNelsonAalen}
\alias{mlr_learners_surv.nelson}
\title{Nelson Aalen Estimator Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Nelson Aalen estimator called from \code{\link[survival:survfit]{survival::survfit()}} in package \CRANpkg{survival}.
}
\details{
The \code{distr} return type is given natively by estimating the cumulative hazard function with \code{\link[survival:survfit]{survival::survfit()}}.\cr
The \code{crank} return type is defined by the expectation of the survival distribution.
}
\section{Construction}{
\preformatted{LearnerSurvNelsonAalen$new()
mlr_learners$get("surv.nelson")
lrn("surv.nelson")
}
}

\references{
Nelson, W. (1969).
Hazard plotting for incomplete failure data.
Journal of Quality Technology, 1, 27–52.
\doi{10.1080/00224065.1969.11980344}.

Nelson, W. (1972).
Theory and applications of hazard plotting for censored failure data.
Technometrics, 14, 945–965.
\doi{10.1080/00401706.1972.10488991}.

Aalen, Odd (1978).
Nonparametric inference for a family of counting processes.
Annals of Statistics, 6(4), 701–726.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
