% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvParametric.R
\docType{data}
\name{LearnerSurvParametric}
\alias{LearnerSurvParametric}
\alias{mlr_learners_surv.parametric}
\title{Fully Parametric Survival Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\description{
A \link{LearnerSurv} for a Fully Parametric model partially implemented in
\code{\link[survival:survreg]{survival::survreg()}} in package \CRANpkg{flexsurv}.
}
\details{
The predict method is based on \code{\link[survival:predict.survreg]{survival::predict.survreg()}} but additionally calculates a survival
distribution using the standard formulae for proportional hazard (PH), accelerated failure time (AFT),
and proportional odds (PO) models. Currently six parameterisations can be assumed for the baseline.
}
\section{Construction}{
\preformatted{LearnerSurvParametric$new()
mlr_learners$get("mlr_learners_surv.parametric")
lrn("mlr_learners_surv.parametric")
}
}

\references{
Kalbfleisch, J. D., Prentice, R. L. (2002).
The Statistical Analysis of Failure Time Data.
John Wiley & Sons.
\doi{10.1002/9781118032985}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
