% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvPenalized.R
\name{LearnerSurvPenalized}
\alias{LearnerSurvPenalized}
\alias{mlr_learners_surv.penalized}
\title{L1 and L2 Penalized Estimation in GLMs Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Calls \code{\link[penalized:penalized]{penalized::penalized()}}.
\itemize{
\item distr is predicted using \code{\link[penalized:predict]{penalized::predict()}}.
\item \code{crank} is predicted as the expectation of the survival distribution, \code{distr}.
}

\code{\link[randomForestSRC:predict.rfsrc]{randomForestSRC::predict.rfsrc()}} returns both cumulative hazard funcion (chf) and survival function (surv)
but uses different estimators to derive these. \code{chf} uses a bootstrapped Nelson-Aalen estimator,
(Ishwaran, 2008) whereas \code{surv} uses a bootstrapped Kaplan-Meier estimator \url{https://kogalur.github.io/randomForestSRC/theory.html}.
The choice of which estimator to use is given by the extra \code{estimator} hyper-parameter,
default is \code{nelson}.
}
\section{Construction}{
\preformatted{LearnerSurvPenalized$new()
mlr_learners$get("surv.penalized")
lrn("surv.penalized")
}
}

\section{Meta Information}{

\itemize{
\item Type: "surv"
\item Predict Types: \verb{distr, crank}
\item Feature Types: \verb{integer, numeric, factor, ordered}
\item Packages: \CRANpkg{penalized} \CRANpkg{distr6}
}
}

\examples{
library(mlr3)
task = tgen("simsurv")$generate(200)
learner = lrn("surv.penalized")
resampling = rsmp("cv", folds = 3)
resample(task, learner, resampling)
}
\references{
Goeman, J. J., L1 penalized estimation in the Cox proportional hazards model.
Biometrical Journal 52(1), 70{84}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}

Other survival learners: 
\code{\link{LearnerSurvBlackboost}},
\code{\link{LearnerSurvCVGlmnet}},
\code{\link{LearnerSurvCoxPH}},
\code{\link{LearnerSurvFlexible}},
\code{\link{LearnerSurvGamboost}},
\code{\link{LearnerSurvGlmboost}},
\code{\link{LearnerSurvGlmnet}},
\code{\link{LearnerSurvKaplan}},
\code{\link{LearnerSurvMboost}},
\code{\link{LearnerSurvNelson}},
\code{\link{LearnerSurvParametric}},
\code{\link{LearnerSurvRandomForestSRC}},
\code{\link{LearnerSurvRanger}},
\code{\link{LearnerSurvRpart}},
\code{\link{LearnerSurvSVM}}
}
\concept{survival learners}
