% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvRandomForestSRC.R
\name{LearnerSurvRandomForestSRC}
\alias{LearnerSurvRandomForestSRC}
\alias{mlr_learners_surv.randomForestSRC}
\title{Survival RandomForestSRC Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
A \link{LearnerSurv} for a survival random forest implemented in \code{\link[randomForestSRC:rfsrc]{randomForestSRC::rfsrc()}} in package \CRANpkg{randomForestSRC}.
}
\details{
The \code{distr} return type is defined from either the cumulative hazard function (chf) or survival function (surv),
predicted by \code{\link[randomForestSRC:predict.rfsrc]{randomForestSRC::predict.rfsrc()}}. Note that these give different results as \code{chf}
uses a bootstrapped Nelson-Aalen estimator (Ishwaran, 2008) whereas \code{surv} uses a
bootstrapped Kaplan-Meier estimator (https://kogalur.github.io/randomForestSRC/theory.html). The choice
of which estimator to use is given by the extra \code{estimator} hyper-parameter, default is Nelson-Aalen.\cr
The \code{crank} return type is defined by the expectation of the survival distribution.
}
\section{Construction}{
\preformatted{LearnerSurvRandomForestSRC$new()
mlr_learners$get("surv.randomForestSRC")
lrn("surv.randomForestSRC")
}
}

\references{
Ishwaran H. and Kogalur U.B. (2019). Fast Unified Random Forests for Survival,
Regression, and Classification (RF-SRC), R package version 2.9.1.

Ishwaran H., Kogalur U.B., Blackstone E.H. and Lauer M.S. (2008). Random survival
forests. Ann. Appl. Statist. 2(3), 841--860.

Breiman, L. (2001).
Random Forests.
Machine Learning 45(1).
\doi{10.1023/A:1010933404324}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
