% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvSVM.R
\name{LearnerSurvSVM}
\alias{LearnerSurvSVM}
\alias{mlr_learners_surv.svm}
\title{Regression, Ranking and Hybrid Support Vector Machine Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Support vector machines predicting either the prognostic index, continuous ranks or a hybrid of the two.
Calls \code{\link[survivalsvm:survivalsvm]{survivalsvm::survivalsvm()}} from package \CRANpkg{survivalsvm}.
}
\details{
Four possible models can be implemented, dependent on the \code{model} parameter. These correspond
to predicting the prognosting index (PI) via regression, \code{regression}; predicting the PI by
imposing a ranking constrant, \code{vanbelle1, vanbelle2}; a hybrid of the two \code{hybrid}. Whilst
\code{regression} is the default, \code{hybrid} may be more efficient in tuning as it can be reduced
to the previous depending on optimal parameters.
}
\section{Construction}{
\preformatted{LearnerSurvSVM$new()
mlr_learners$get("surv.svm")
lrn("surv.svm")
}
}

\examples{
library(mlr3)
task = tgen("simsurv")$generate(200)
learner = lrn("surv.svm")
resampling = rsmp("cv", folds = 3)
resample(task, learner, resampling)
}
\references{
Van Belle, V., Pelcmans, K., Van Huffel S. and Suykens J. A.K. (2011a).
Improved performance on high-dimensional survival data by application of Survival-SVM.
Bioinformatics, 27, 87-94.

Van Belle, V., Pelcmans, K., Van Huffel S. and Suykens J. A.K. (2011b).
Support vector methods for survival analysis: a comparison between ranking and regression approaches.
Artificial Intelligence in Medicine, 53, 107-118.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
