% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegrIntLogloss.R
\name{MeasureRegrIntLogloss}
\alias{MeasureRegrIntLogloss}
\alias{mlr_measures_regr.intlogloss}
\title{Integrated Regression Log loss Regression Measure}
\description{
If \code{integrated == FALSE} then the sample mean is taken for the single specified \code{times}, \eqn{t^*}{t*}, and the returned
score is given by
\deqn{L(S,t|t^*) = \frac{1}{N} \sum_{i=1}^N L(S_i,t_i|t^*)}{L(S,t|t*) = 1/N \sum_i^N L(S_i,t_i|t*)}
where \eqn{N} is the number of observations, \eqn{S_i} is the predicted survival function for
individual \eqn{i} and \eqn{t_i} is their true survival time.

If \code{integrated == TRUE} then an approximation to integration is made by either taking the sample
mean over all \eqn{T} unique time-points (\code{method == 1}), or by taking a mean weighted by the difference
between time-points (\code{method == 2}). Then the sample mean is taken over all \eqn{N} observations.
\deqn{L(S) = \frac{1}{NT} \sum_{i=1}^N \sum_{j=1}^T L(S_i,t_i|t^*_j)}{L(S) = 1/(NT) \sum_i^N \sum_j^T L(S_i,t_i|t*_j)}

Calculates the integrated logarithmic (log), loss, aka integrated cross entropy.

For an individual with outcome \eqn{Y} and predicted CDF, \eqn{F}, the
integrated log loss is given by
\deqn{L(F, y) = - \int I(Y \le y)log(F(y)) + I(Y > y)log(1-F(y)) dy}
}
\section{Dictionary}{

This \link[mlr3:Measure]{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link[mlr3:mlr_measures]{mlr_measures} or with the associated sugar function \link[mlr3:msr]{msr()}:\preformatted{MeasureRegrIntLogloss$new(eps = 1e-15)
mlr_measures$get("regr.intlogloss")
msr("regr.intlogloss")
}
}

\section{Meta Information}{

\itemize{
\item Type: \code{"regr"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{distr}
}
}

\seealso{
Other regression measures: 
\code{\link{MeasureRegrBrierSE}},
\code{\link{MeasureRegrIntLoglossSE}},
\code{\link{MeasureRegrLogloss}}

Other Probabilistic regression measures: 
\code{\link{MeasureRegrBrierSE}},
\code{\link{MeasureRegrIntLoglossSE}},
\code{\link{MeasureRegrLogloss}}
}
\concept{Probabilistic regression measures}
\concept{regression measures}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureRegr]{mlr3::MeasureRegr}} -> \code{\link[mlr3proba:MeasureRegrIntegrated]{mlr3proba::MeasureRegrIntegrated}} -> \code{MeasureRegrIntLogloss}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{eps}}{Returns \code{eps} parameter, see \code{initialize}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureRegrIntLogloss$new()}}
\item \href{#method-clone}{\code{MeasureRegrIntLogloss$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureRegrIntLogloss$new(integrated = TRUE, points, eps = 1e-15, method = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{integrated}}{(\code{logical(1)})\cr
If \code{TRUE} (default), returns the integrated score; otherwise, not integrated.}

\item{\code{points}}{(\code{numeric()})\cr
If \code{integrate == TRUE} then a vector of points over which to integrate the score.
If \code{integrate == FALSE} then a single point at which to return the score.}

\item{\code{eps}}{(\code{numeric(1)})\cr
Very small number to set zero-valued predicted probabilities to in order to prevent errors
in log(0) calculation.}

\item{\code{method}}{(\code{integer(1)})\cr
If \code{integrate == TRUE} selects the integration weighting method.
\code{method == 1} corresponds to weighting each time-point equally and taking the mean score over
discrete time-points. \code{method == 2} corresponds to calculating a mean weighted by the difference
between time-points. \code{method == 2} is default to be in line with other packages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureRegrIntLogloss$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
