% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurv.R
\name{MeasureSurv}
\alias{MeasureSurv}
\title{Survival Measure}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Measure]{mlr3::Measure}.}
\description{
This measure specializes \link[mlr3:Measure]{mlr3::Measure} for survival problems.
Predefined measures can be found in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link[mlr3:mlr_measures]{mlr3::mlr_measures}.
}
\section{Construction}{
\preformatted{MeasureSurv$new(id, range, minimize, aggregator = NULL,
                properties = character(), predict_type = "distr",
                task_properties = character(), packages = character())
}

For a description of the arguments, see \link[mlr3:Measure]{mlr3::Measure}.
The \code{task_type} is set to \code{"surv"}.
Possible values for \code{predict_type} are \code{"distr"}, \code{"lp"}, \code{"crank"}, and \code{"response"}.
}

\section{Fields}{

See \link{Measure}.
}

\section{Methods}{

See \link{Measure}.
}

\seealso{
Example survival measures: \code{\link[=MeasureSurvGraf]{surv.graf}},
\code{\link[=MeasureSurvHarrellC]{surv.harrellC}}.

Other Measure: 
\code{\link{MeasureDens}}
}
\concept{Measure}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureSurv}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureSurv$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureSurv$new(
  id,
  range,
  minimize = NA,
  aggregator = NULL,
  properties = character(),
  predict_type = "distr",
  task_properties = character(),
  packages = character()
)}\if{html}{\out{</div>}}
}

}
}
