% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeaureSurvAUC.R
\name{MeasureSurvAUC}
\alias{MeasureSurvAUC}
\title{Abstract Class for AUC-type Survival Measures}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
This class is abstract and should not be constructed directly.
}
\section{Fields}{

See \link{MeasureSurv} as well as
\itemize{
\item integrated :: logical(1) \cr Should the integrated AUC be returned?
\item times :: numeric() \cr Either the times over which to integrate the AUC or time-points for which (non-integrated) AUC should be returned.
}
}

