% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvBeggC.R
\name{MeasureSurvBeggC}
\alias{MeasureSurvBeggC}
\alias{mlr_measures_surv.beggC}
\title{Begg's C-Index}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:BeggC]{survAUC::BeggC()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes Cox PH model specification.
}
\section{Construction}{
\preformatted{MeasureSurvBeggC$new()
mlr_measures$get("surv.beggC")
msr("surv.beggC")
}
}

\references{
Begg, B. C., L. D. Craemer, E. S. Venkatraman and J. Rosai (2000).
Comparing tumor staging and grading systems: a case study and a review of the issues, using thymoma as a model.
Statistics in Medicine 19, 1997–2014.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
