% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvBrier.R
\docType{data}
\name{MeasureSurvBrier}
\alias{MeasureSurvBrier}
\alias{mlr_measures_surv.brier}
\title{Brier Score}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calculates the brier score, aka squared loss.
}
\section{Construction}{
\preformatted{MeasureSurvBrier$new()
mlr_measures$get("surv.brier")
msr("surv.brier")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
\keyword{datasets}
