% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvChamblessAUC.R
\name{MeasureSurvChamblessAUC}
\alias{MeasureSurvChamblessAUC}
\alias{mlr_measures_surv.chamblessAUC}
\title{Chambless and Diao's AUC}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:AUC.cd]{survAUC::AUC.cd()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes Cox PH model specification.
}
\section{Construction}{
\preformatted{MeasureSurvChamblessAUC$new()
mlr_measures$get("surv.chamblessAUC")
msr("surv.chamblessAUC")
}
}

\references{
Chambless, L. E. and G. Diao (2006).
Estimation of time-dependent area under the ROC curve for long-term risk prediction.
Statistics in Medicine 25, 3474–3486.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
