% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvGonenHellersC.R
\name{MeasureSurvGonensC}
\alias{MeasureSurvGonensC}
\alias{mlr_measures_surv.gonensC}
\title{Gonen and Heller's C-Index}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:GHCI]{survAUC::GHCI()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes Cox PH model specification.
}
\section{Construction}{
\preformatted{MeasureSurvGonensC$new()
mlr_measures$get("surv.gonensC")
msr("surv.gonensC")
}
}

\references{
Gonen, M. and G. Heller (2005).
Concordance probability and discriminatory power in proportional hazards regression.
Biometrika 92, 965–970.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
