% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvGrafSE.R
\name{MeasureSurvGrafSE}
\alias{MeasureSurvGrafSE}
\alias{mlr_measures_surv.grafSE}
\title{Standard Error of Integrated Graf Score Survival Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{


Calculates the standard error of \link{MeasureSurvGraf}.

The standard error is approximated using Binomial approxiation. For \eqn{N} observations in the
test set, the standard eror is given by
\deqn{IGS_SE = sd(IGS(S))/sqrt(N)}
}
\section{Construction}{
\preformatted{MeasureSurvGrafSE$new(integrated = TRUE, times)
mlr_measures$get("surv.grafSE")
msr("surv.grafSE")
}
\itemize{
\item \code{integrated} :: \code{logical(1)} \cr If \code{TRUE} (default), returns the integrated score, otherwise returns the score corresponding to the \code{times} argument. \cr
\item \code{times} :: \code{vector()} \cr Numeric vector of times where to evaluate the score (provide only one for best performance). If missing then returns score at all unique time-points in test set. \cr
}
}

\section{Meta Information}{

\itemize{
\item Type: \code{"surv"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{distr}
}
}

\references{
Graf, E., Schmoor, C., Sauerbrei, W. and Schumacher, M. (1999).\cr
Assessment and comparison of prognostic classification schemes for survival data.\cr
Statistics in Medicine, 18(17), 2529-2545.\cr
\doi{10.1002/(SICI)1097-0258(19990915/30)18:17/18<2529::AID-SIM274>3.0.CO;2-5}
}
\seealso{
Other survival measures: 
\code{\link{MeasureSurvBeggC}},
\code{\link{MeasureSurvChamblessAUC}},
\code{\link{MeasureSurvGonenC}},
\code{\link{MeasureSurvGraf}},
\code{\link{MeasureSurvHarrellC}},
\code{\link{MeasureSurvHungAUC}},
\code{\link{MeasureSurvLoglossIntSE}},
\code{\link{MeasureSurvLoglossInt}},
\code{\link{MeasureSurvLoglossSE}},
\code{\link{MeasureSurvLogloss}},
\code{\link{MeasureSurvNagelkR2}},
\code{\link{MeasureSurvOQuigleyR2}},
\code{\link{MeasureSurvSongAUC}},
\code{\link{MeasureSurvSongTNR}},
\code{\link{MeasureSurvSongTPR}},
\code{\link{MeasureSurvUnoAUC}},
\code{\link{MeasureSurvUnoC}},
\code{\link{MeasureSurvUnoTNR}},
\code{\link{MeasureSurvUnoTPR}},
\code{\link{MeasureSurvXuR2}}

Other Probabilistic survival measures: 
\code{\link{MeasureSurvGraf}},
\code{\link{MeasureSurvLoglossIntSE}},
\code{\link{MeasureSurvLoglossInt}},
\code{\link{MeasureSurvLoglossSE}},
\code{\link{MeasureSurvLogloss}}
}
\concept{Probabilistic survival measures}
\concept{survival measures}
