% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvHarrellsC.R
\name{MeasureSurvHarrellsC}
\alias{MeasureSurvHarrellsC}
\alias{mlr_measures_surv.harrellsC}
\title{Harrell's C-Index}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calculates Harrell's C, equivalent to the Fortran implementation in \CRANpkg{Hmisc}.
}
\details{
Requires \code{crank} \code{predict_type}.
}
\section{Construction}{
\preformatted{MeasureSurvHarrellsC$new()
mlr_measures$get("surv.harrellsC")
msr("surv.harrellsC")
}
}

\references{
Harrell, F. E., R. M. Califf, D. B. Pryor, K. L. Lee and R. A. Rosati (1982).
Evaluating the yield of medical tests.
Journal of the American Medical Association. 247. 2543–2546.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
