% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvHungAUC.R
\name{MeasureSurvHungAUC}
\alias{MeasureSurvHungAUC}
\alias{mlr_measures_surv.hungAUC}
\title{Hung and Chiang's AUC}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:AUC.hc]{survAUC::AUC.hc()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes random censoring.
}
\section{Construction}{
\preformatted{MeasureSurvHungAUC$new()
mlr_measures$get("surv.hungAUC")
msr("surv.hungAUC")
}
}

\references{
Hung, H. and C.-T. Chiang (2010).
Estimation methods for time-dependent AUC models with survival data.
Canadian Journal of Statistics 38, 8–26.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
