% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvIntegrated.R
\name{MeasureSurvIntegrated}
\alias{MeasureSurvIntegrated}
\title{Abstract Class for Integrated Survival Measures}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
This class is abstract and should not be constructed directly.
}
\section{Fields}{

See \link{MeasureSurv} as well as
\itemize{
\item integrated :: logical(1) \cr Should the integrated score be returned?
\item times :: numeric() \cr Either the times over which to integrate the score or a \emph{single} time-point for which the (non-integrated) score should be returned.
}
}

