% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvLogloss.R
\docType{data}
\name{MeasureSurvLogloss}
\alias{MeasureSurvLogloss}
\alias{mlr_measures_surv.logloss}
\title{Logloss}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calculates the cross-entropy, or logarithmic, loss.
}
\section{Construction}{
\preformatted{MeasureSurvLogloss$new()
mlr_measures$get("surv.logloss")
msr("surv.logloss")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
\keyword{datasets}
