% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvIntLogloss.R
\name{MeasureSurvLoglossInt}
\alias{MeasureSurvLoglossInt}
\alias{mlr_measures_surv.loglossint}
\title{Integrated Logloss}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calculates the integrated cross-entropy, or logarithmic, loss.
}
\details{
The integrated logloss, in the context of probabilistic predictions, is defined as the classical
logloss for a given individual, integrated over all time-points.
}
\section{Construction}{
\preformatted{MeasureSurvLoglossInt$new()
mlr_measures$get("surv.loglossint")
msr("surv.loglossint")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
