% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvOQuigleyR2.R
\name{MeasureSurvOQuigleyR2}
\alias{MeasureSurvOQuigleyR2}
\alias{mlr_measures_surv.oquigleyR2}
\title{O'Quigley, Xu, and Stare's R2 Coefficient}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:OXS]{survAUC::OXS()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes Cox PH model specification.
}
\section{Construction}{
\preformatted{MeasureSurvOQuigleyR2$new()
mlr_measures$get("surv.oquigleyR2")
msr("surv.oquigleyR2")
}
}

\references{
O'Quigley, J., R. Xu, and J. Stare (2005).
Explained randomness in proportional hazards models.
Statistics in Medicine 24, 479–489.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
