% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvSongAUC.R
\name{MeasureSurvSongAUC}
\alias{MeasureSurvSongAUC}
\alias{mlr_measures_surv.songAUC}
\title{Song and Zhou's AUC}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:AUC.sh]{survAUC::AUC.sh()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes Cox PH model specification.
}
\section{Construction}{
\preformatted{MeasureSurvSongAUC$new()
mlr_measures$get("surv.songAUC")
msr("surv.songAUC")
}
}

\references{
Song, X. and X.-H. Zhou (2008).
A semiparametric approach for the covariate specific ROC curve with survival outcome.
Statistica Sinica 18, 947–965.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
