% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvUnosC.R
\name{MeasureSurvUnosC}
\alias{MeasureSurvUnosC}
\alias{mlr_measures_surv.unosC}
\title{Uno's C-Index}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:UnoC]{survAUC::UnoC()}}.
}
\details{
Requires \code{crank} \code{predict_type}. \cr
Assumes random censoring assumption holds.
}
\section{Construction}{
\preformatted{MeasureSurvUnosC$new()
mlr_measures$get("surv.unosC")
msr("surv.unosC")
}
}

\references{
Uno, H., T. Cai T, M. J. Pencina, R. B. D'Agostino and W. L. Wei. (2011).
On the C-statistics for evaluating overall adequacy of risk prediction procedures with censored
survival data.
Statistics in Medicine. 30(10): 1105–1117.
\doi{10.1002/sim.4154}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
