% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvXusR2.R
\name{MeasureSurvXusR2}
\alias{MeasureSurvXusR2}
\alias{mlr_measures_surv.xusR2}
\title{Xu and O'Quigley's R2 Coefficient}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:XO]{survAUC::XO()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes Cox PH model specification.
}
\section{Construction}{
\preformatted{MeasureSurvXusR2$new()
mlr_measures$get("surv.xusR2")
msr("surv.xusR2")
}
}

\references{
Xu, R. and J. O'Quigley (1999).
A measure of dependence for proportional hazards models.
Journal of Nonparametric Statistics 12, 83–107.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
