% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureCompRisksAUC.R
\name{mlr_measures_cmprsk.auc}
\alias{mlr_measures_cmprsk.auc}
\alias{MeasureCompRisksAUC}
\title{Blanche's AUC Competing Risks Measure}
\description{
Calculates the cause-specific ROC-AUC(t) at a \strong{specific time point},
see Blanche et al. (2013).
Can also return the mean AUC(t) over all competing causes.
}
\details{
Calls \code{\link[riskRegression:Score]{riskRegression::Score()}} with:
\itemize{
\item \code{metric = "auc"}
\item \code{cens.method = "ipcw"}
\item \code{cens.model = "km"}
}

Notes on the \code{riskRegression} implementation:
\enumerate{
\item IPCW weights are estimated using the \strong{test data only}.
\item No extrapolation is supported: if \code{time_horizon} exceeds the maximum observed
time on the test data, an error is thrown.
}
}
\section{Dictionary}{

This \link[mlr3:Measure]{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link[mlr3:mlr_measures]{mlr_measures} or with the associated sugar function \link[mlr3:mlr_sugar]{msr()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{MeasureCompRisksAUC$new()
mlr_measures$get("cmprsk.auc")
msr("cmprsk.auc")
}\if{html}{\out{</div>}}
}

\section{Parameters}{
\tabular{llll}{
   Id \tab Type \tab Default \tab Range \cr
   cause \tab integer \tab 1 \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   time_horizon \tab numeric \tab NULL \tab \eqn{[0, \infty)}{[0, Inf)} \cr
}
}

\section{Meta Information}{

\itemize{
\item Type: \code{"cmprsk"}
\item Range: \eqn{[0, 1]}{[0, 1]}
\item Minimize: \code{FALSE}
\item Required prediction: \code{cif}
}
}

\section{Parameter details}{

\itemize{
\item \code{cause} (\code{numeric(1)})\cr
Integer number indicating which cause to use (Default: \code{1}).
If \code{"mean"}, then the mean AUC(t) over all causes is returned.
\item \code{time_horizon} (\code{numeric(1)})\cr
Single time point at which to return the score.
If \code{NULL}, the \strong{median time point} from the test set is used.
}
}

\examples{
\dontshow{if (mlr3misc::require_namespaces("riskRegression", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mlr3)

# Define the Learner
learner = lrn("cmprsk.aalen")
learner

# Define a Task
task = tsk("pbc")

# Stratification based on event
task$set_col_roles(cols = "status", add_to = "stratum")

# Create train and test set
part = partition(task)

# Train the learner on the training set
learner$train(task, row_ids = part$train)
learner$model

# Make predictions for the test set
predictions = learner$predict(task, row_ids = part$test)
predictions

# Score the predictions

## AUC(t = 100), 1st cause
predictions$score(msr("cmprsk.auc", cause = 1, time_horizon = 100))
## AUC(t = 100), 2nd cause
predictions$score(msr("cmprsk.auc", cause = 2, time_horizon = 100))
## AUC(t = 100), summary score across causes
predictions$score(msr("cmprsk.auc", cause = "mean", time_horizon = 100))

\dontshow{\}) # examplesIf}
}
\references{
Blanche, Paul, Dartigues, Francois J, Jacqmin-Gadda, Helene (2013).
\dQuote{Estimating and comparing time-dependent areas under receiver operating characteristic curves for censored event times with competing risks.}
\emph{Statistics in Medicine}, \bold{32}(30), 5381--5397.
ISSN 1097-0258, \doi{10.1002/SIM.5958}, \url{https://onlinelibrary.wiley.com/doi/10.1002/sim.5958}.
}
\concept{Competing risks measures}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3proba:MeasureCompRisks]{mlr3proba::MeasureCompRisks}} -> \code{MeasureCompRisksAUC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureCompRisksAUC-new}{\code{MeasureCompRisksAUC$new()}}
\item \href{#method-MeasureCompRisksAUC-clone}{\code{MeasureCompRisksAUC$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureCompRisksAUC-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureCompRisksAUC-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureCompRisksAUC$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureCompRisksAUC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureCompRisksAUC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureCompRisksAUC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
