% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpDistrCompositor.R
\docType{data}
\name{mlr_pipeops_distrcompose}
\alias{mlr_pipeops_distrcompose}
\alias{PipeOpDistrCompositor}
\title{PipeOpDistrCompositor}
\format{\code{\link{R6Class}} inheriting from \code{\link{PipeOpCompositor}}/\code{\link{PipeOp}}.}
\description{
Create a survival distribution prediction from two survival learner \code{\link[mlr3:Prediction]{Prediction}}s,
where one predicts a relative risk, and the other estimates the baseline hazard.

The resulting \code{"distr"} prediction is a combination of the the incoming \code{"risk"} or \code{"lp"} prediction
and incoming \code{"distr"} prediction. The \code{model_form} parameter determines what form the composed
distribution should assume, this is one of proportional hazards (\code{ph}), accelerated failure time (\code{aft}),
or proportional odds (\code{po}). The hazard functions of these are defined respectively by
\deqn{PH: h(t) = h0(t)exp(lp)}
}
\section{Construction}{
\preformatted{PipeOpDistrCompositor$new(id = "distrcompose", model_form = "aft", param_vals = list())
}
\itemize{
\item \code{id} :: \code{character(1)}
Identifier of the resulting  object, default \code{"distrcompose"}.
\item \code{model_form} :: \code{character(1)}
Form of the composed distribution, one of: \code{aft} (accelerated failure time), \code{ph}, \code{po}.
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default \code{list()}.
}
}

\section{Input and Output Channels}{

Input and output channels are inherited from \code{\link{PipeOpCompositor}}. Instead of a \code{\link[mlr3:Prediction]{Prediction}}, a \code{\link[mlr3:PredictionRegr]{PredictionRegr}}
is used as input and output during prediction.
}

\section{State}{

The \code{$state} is left empty (\code{list()}).
}

\section{Parameters}{

The parameters are the parameters inherited from the \code{\link{PipeOpCompositor}}.
}

\section{Internals}{

Inherits from \code{\link{PipeOpCompositor}} by implementing the \code{private$weighted_avg_predictions()} method.
}

\section{Fields}{

Only fields inherited from \code{\link{PipeOpCompositor}}/\code{\link{PipeOp}}.
}

\section{Methods}{

Only methods inherited from \code{\link{PipeOpCompositor}}/\code{\link{PipeOp}}.
}

\seealso{
Other PipeOps: \code{\link{PipeOpCompositor}}

Other Ensembles: \code{\link{PipeOpCompositor}}
}
\concept{Ensembles}
\concept{PipeOps}
\keyword{datasets}
