% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpPredClassifSurv.R
\name{mlr_pipeops_trafopred_classifsurv}
\alias{mlr_pipeops_trafopred_classifsurv}
\alias{PipeOpPredClassifSurv}
\title{PipeOpPredClassifSurv}
\description{
Transform \link{PredictionClassif} to \link{PredictionSurv} by converting
event probabilities of a pseudo status variable (discrete time hazards)
to survival probabilities.
}
\section{Input and Output Channels}{

Input and output channels are inherited from \link[mlr3pipelines:PipeOp]{PipeOp}.

The output is the input \link{PredictionClassif} transformed to a \link{PredictionSurv}.
}

\examples{
\dontrun{
if (requireNamespace("mlr3pipelines", quietly = TRUE)) {
  library(mlr3)
  library(mlr3pipelines)

  task = tsk("rats")

  if (requireNamespace("mlr3learners", quietly = TRUE)) {
    library(mlr3learners)
    po_tasktoclassif = po("trafotask_survclassif")
    po_tasktoclassif$train(list(task))
    task_classif = po_tasktoclassif$predict(list(task))[[1]]
    trafo_data = po_tasktoclassif$predict(list(task))[[2]]

    learner = lrn("classif.log_reg", predict_type = "prob")
    learner$train(task_classif)
    pred = learner$predict(task_classif)

    po_predtosurv = po("trafopred_classifsurv")
    po_predtosurv$train(list(pred, trafo_data))
    po_predtosurv$predict(list(pred, trafo_data))
  }
}
}
}
\seealso{
Other PipeOps: 
\code{\link{PipeOpPredTransformer}},
\code{\link{PipeOpTaskTransformer}},
\code{\link{PipeOpTransformer}},
\code{\link{mlr_pipeops_survavg}},
\code{\link{mlr_pipeops_trafopred_regrsurv}},
\code{\link{mlr_pipeops_trafopred_survregr}},
\code{\link{mlr_pipeops_trafotask_regrsurv}},
\code{\link{mlr_pipeops_trafotask_survclassif}},
\code{\link{mlr_pipeops_trafotask_survregr}}

Other Transformation PipeOps: 
\code{\link{mlr_pipeops_trafopred_regrsurv}},
\code{\link{mlr_pipeops_trafopred_survregr}},
\code{\link{mlr_pipeops_trafotask_regrsurv}},
\code{\link{mlr_pipeops_trafotask_survclassif}},
\code{\link{mlr_pipeops_trafotask_survregr}}
}
\concept{PipeOps}
\concept{Transformation PipeOps}
\section{Super class}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{PipeOpPredClassifSurv}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpPredClassifSurv-new}{\code{PipeOpPredClassifSurv$new()}}
\item \href{#method-PipeOpPredClassifSurv-clone}{\code{PipeOpPredClassifSurv$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpPredClassifSurv-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpPredClassifSurv-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpPredClassifSurv$new(id = "trafopred_classifsurv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(character(1))\cr
Identifier of the resulting object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpPredClassifSurv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpPredClassifSurv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpPredClassifSurv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
