% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskDens_X.R
\name{mlr_tasks_faithful}
\alias{mlr_tasks_faithful}
\title{Old Faithful Eruptions Density Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskDens}.
}
\description{
A density task for the \link{faithful} data set.
}
\details{
Only the \code{eruptions} column is kept in this task.
}
\section{Construction}{
\preformatted{mlr3::mlr_tasks$get("faithful")
mlr3::tsk("faithful")
}
}

