% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv_X.R
\name{mlr_tasks_gbcs}
\alias{mlr_tasks_gbcs}
\title{German Breast Cancer Study Survival Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.
}
\description{
A survival task for the \link{gbcs} data set.
}
\details{
Column "id" and all date columns removed, as well as "rectime" and "censrec".
Target is time to death.
}
\section{Construction}{
\preformatted{mlr3::mlr_tasks$get("gbcs")
mlr3::tsk("gbcs")
}
}

