% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv_X.R
\name{mlr_tasks_grace}
\alias{mlr_tasks_grace}
\title{GRACE 1000 Survival Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.
}
\description{
A survival task for the \link{grace} data set.
}
\details{
Column "id" is removed. Columns "days" and "death" renamed to "time" and "status" resp.
}
\section{Construction}{
\preformatted{mlr3::mlr_tasks$get("grace")
mlr3::tsk("grace")
}
}

