% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskDens_X.R
\name{mlr_tasks_precip}
\alias{mlr_tasks_precip}
\title{Annual Precipitation Density Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskDens}.
}
\description{
A density task for the \link{precip} data set.
}
\section{Construction}{
\preformatted{mlr3::mlr_tasks$get("precip")
mlr3::tsk("precip")
}
}

