% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv_rats.R
\name{mlr_tasks_rats}
\alias{mlr_tasks_rats}
\title{Rats Survival Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.}
\description{
A survival task for the \link[survival:rats]{survival::rats} data set.
Column "sex" has been converted to a factor, all others have been converted to integer.
}
\section{Construction}{
\preformatted{mlr3::mlr_tasks$get("rats")
mlr3::tsk("rats")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Task]{Tasks}: \link[mlr3:mlr_tasks]{mlr3::mlr_tasks}
}
