% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv_X.R
\name{mlr_tasks_whas}
\alias{mlr_tasks_whas}
\title{Worcester Heart Attack Study (WHAS) Survival Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.
}
\description{
A survival task for the \link{whas} data set.
}
\details{
Columns "id", "yrgrp", and "dstat" are removed so target is status at last follow-up.
Column "sex" renamed to "sexF", "lenfol" to "time", and "fstat" to "status".
}
\section{Construction}{
\preformatted{mlr3::mlr_tasks$get("whas")
mlr3::tsk("whas")
}
}

