% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_survbagging}
\alias{pipeline_survbagging}
\title{Bag Survival Predictions}
\usage{
pipeline_survbagging(
  learner,
  iterations = 10,
  frac = 0.7,
  weights = 1,
  graph_learner = TRUE
)
}
\arguments{
\item{learners}{\code{(list())} \cr
List of \link{LearnerSurv}s to average.}

\item{param_vals}{\code{(list())} \cr
Parameters, including weights, to pass to \link{PipeOpSurvAvg}.}
}
\value{
\link[mlr3pipelines:GraphLearner]{mlr3pipelines::GraphLearner}
}
\description{
This is a bagging graph based around the \link{PipeOpSurvAvg} pipe operation, which first
subsamples the data and then fits the same learner multiple times and aggregates the results.
}
\details{
For full details see \link{PipeOpSurvAvg}.
}
\examples{
\dontrun{
library("mlr3")

task = tgen("simsurv")$generate(5)
avg = surv_averager(
  learners = lrns(c("surv.kaplan", "surv.coxph")),
  param_vals = list(weights = c(0.1, 0.9))
 )
avg$train(task)$predict(task)
}
}
