% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{plot.PredictionSurv}
\alias{plot.PredictionSurv}
\title{Calibration Plot for \code{PredictionSurv} Objects}
\usage{
\method{plot}{PredictionSurv}(x, task, row_ids = NULL, times = NULL, ...)
}
\arguments{
\item{x}{(\link{PredictionSurv})}

\item{task}{(\link{TaskSurv}) \cr
For Kaplan-Meier learner, passed to \verb{$predict}.}

\item{row_ids}{(\code{integer()}) \cr
For Kaplan-Meier learner, passed to \verb{$predict}.}

\item{times}{(\code{numeric()}) \cr
Times for plotting over, if \code{NULL} uses all times from \code{task}.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Compares the average predicted survival distribution to a Kaplan-Meier estimate.
}
\examples{
\donttest{
learn = lrn("surv.coxph")
task = tsk("rats")
p = learn$train(task, row_ids = 1:100)$predict(task, row_ids = 101:200)
plot(p, task)
}

}
