% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCompRisks.R
\name{PredictionCompRisks}
\alias{PredictionCompRisks}
\title{Prediction Object for Competing Risks}
\description{
This object stores the predictions returned by a learner of class \link{LearnerCompRisks}.

The \code{task_type} is set to \code{"cmprsk"}.

For accessing survival and hazard functions, as well as other complex methods
from a \link{LearnerCompRisks} object is not possible atm.
}
\examples{
library(mlr3)
task = tsk("pbc")
learner = lrn("cmprsk.aalen")
part = partition(task)
p = learner$train(task, part$train)$predict(task, part$test)
p

# CIF list: 1 matrix (obs x times) per competing event
names(p$cif) # competing events
# CIF matrix for competing event 1 (first 5 test observations and 20 time points)
p$cif[["1"]][1:5, 1:20]
# CIF matrix for competing event 2 (first 5 test observations and 20 time points)
p$cif[["2"]][1:5, 1:20]

# data.table conversion
tab = as.data.table(p)
tab$CIF[[1]] # for first test observation, list of CIF vectors

}
\seealso{
Other Prediction: 
\code{\link{PredictionDens}},
\code{\link{PredictionSurv}}
}
\concept{Prediction}
\section{Super class}{
\code{\link[mlr3:Prediction]{mlr3::Prediction}} -> \code{PredictionCompRisks}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{truth}}{(\code{Surv})\cr
True (observed) outcome.}

\item{\code{cif}}{(\code{list()})\cr
Access the stored CIFs.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionCompRisks-new}{\code{PredictionCompRisks$new()}}
\item \href{#method-PredictionCompRisks-clone}{\code{PredictionCompRisks$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="filter"><a href='../../mlr3/html/Prediction.html#method-Prediction-filter'><code>mlr3::Prediction$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="format"><a href='../../mlr3/html/Prediction.html#method-Prediction-format'><code>mlr3::Prediction$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="help"><a href='../../mlr3/html/Prediction.html#method-Prediction-help'><code>mlr3::Prediction$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="obs_loss"><a href='../../mlr3/html/Prediction.html#method-Prediction-obs_loss'><code>mlr3::Prediction$obs_loss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="print"><a href='../../mlr3/html/Prediction.html#method-Prediction-print'><code>mlr3::Prediction$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="score"><a href='../../mlr3/html/Prediction.html#method-Prediction-score'><code>mlr3::Prediction$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionCompRisks-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionCompRisks-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionCompRisks$new(
  task = NULL,
  row_ids = task$row_ids,
  truth = task$truth(),
  cif = NULL,
  check = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link{TaskSurv})\cr
Task, used to extract defaults for \code{row_ids} and \code{truth}.}

\item{\code{row_ids}}{(\code{integer()})\cr
Row ids of the predicted observations, i.e. the row ids of the test set.}

\item{\code{truth}}{(\code{survival::Surv()})\cr
True (observed) response.}

\item{\code{cif}}{(\code{list()})\cr
A \code{list} of two or more \code{matrix} objects.
Each matrix represents a different competing event and it stores the
\strong{Cumulative Incidence function} for each test observation.
In each matrix, rows represent observations and columns time points.
The names of the \code{list} must correspond to the competing event names
(\code{task$cmp_events}).}

\item{\code{check}}{(\code{logical(1)})\cr
If \code{TRUE}, performs argument checks and predict type conversions.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \code{cif} input currently is a list of CIF matrices.
In the future the \code{cif} field should coerce this to a (TODO) CIF container
to get all survival and hazard functions using interpolation.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionCompRisks-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionCompRisks-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionCompRisks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
