% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionDensity.R
\name{PredictionDensity}
\alias{PredictionDensity}
\title{Prediction Object for Density}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Prediction]{mlr3::Prediction}.}
\description{
This object stores the predictions returned by a learner of class \link{LearnerDensity}.

The \code{task_type} is set to \code{"density"}.
}
\section{Construction}{
\preformatted{p = PredictionDensity$new(task = NULL, row_ids = task$row_ids, truth = task$truth(),
pdf = pdf, cdf = cdf)
}
\itemize{
\item \code{task} :: \link{TaskDensity}\cr
Task, used to extract defaults for \code{row_ids} and \code{truth}.
\item \code{row_ids} :: (\code{integer()} | \code{character()})\cr
Row ids of the task. Per default, these are extracted from the \code{task}.
\item \code{truth} :: \code{numeric()}\cr
Observed sample. Per default, these are extracted from the \code{task}.
\item \code{pdf} :: \code{numeric()}\cr
Probability density function evaluated at the given points in the test set.
\item \code{cdf} :: \code{numeric()}\cr
Cumulative distribution function evaluated at the given points in the test set.
}
}

\section{Fields}{

See \link[mlr3:Prediction]{mlr3::Prediction}.

The field \code{task_type} is set to \code{"density"}.
}

\examples{
library(mlr3)
task = mlr_tasks$get("precip")
learner = mlr_learners$get("density.hist")
p = learner$train(task)$predict(task)
head(as.data.table(p))
}
\seealso{
Other Prediction: 
\code{\link{PredictionSurv}}
}
\concept{Prediction}
