% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breslow.R
\name{surv_breslow}
\alias{surv_breslow}
\title{Survival probabilities using Breslow's estimator}
\usage{
surv_breslow(times, status, lp_train, lp_test, eval_times = NULL)
}
\arguments{
\item{times}{(\code{numeric()})\cr Vector of times (train set).}

\item{status}{(\code{numeric()})\cr Vector of status indicators (train set).
For each observation in the train set, this should be 0 (alive/censored) or
1 (dead).}

\item{lp_train}{(\code{numeric()})\cr Vector of linear predictors (train set).
These are the relative score predictions (\eqn{lp = \hat{\beta}X_{train}})
from a proportional hazards model on the train set.}

\item{lp_test}{(\code{numeric()})\cr Vector of linear predictors (test set).
These are the relative score predictions (\eqn{lp = \hat{\beta}X_{test}})
from a proportional hazards model on the test set.}

\item{eval_times}{(\code{numeric()})\cr Vector of times to compute survival
probabilities. If \code{NULL} (default), the unique and sorted \code{times} from the
train set will be used, otherwise the unique and sorted \code{eval_times}.}
}
\value{
a survival probability \code{matrix} (obs x times). Number of columns is
equal to \code{eval_times} and number of rows is equal to the number of test
observations (i.e. the length of the \code{lp_test} vector).
}
\description{
Helper function to compose a survival distribution from the relative risk
predictions (\code{lp}) of a \strong{proportional hazards} model (e.g. a Cox-type model).
}
\details{
We estimate the survival probability of individual \eqn{i} (from the test set),
at time point \eqn{t} as follows:
\deqn{S_i(t) = e^{-H_i(t)} = e^{-\hat{H}_0(t) \times e^{lp_i}}}

where:
\itemize{
\item \eqn{H_i(t)} is the cumulative hazard function for individual \eqn{i}
\item \eqn{\hat{H}_0(t)} is Breslow's estimator for the cumulative baseline
hazard. Estimation requires the training set's \code{times} and \code{status} as well
the risk predictions (\code{lp_train}), see more details in \link{.cbhaz_breslow})
\item \eqn{lp_i} is the risk prediction (linear predictor) of individual \eqn{i}
on the test set.
}
}
\examples{
task = tsk("rats")
part = partition(task, ratio = 0.8)

learner = lrn("surv.coxph")
learner$train(task, part$train)
p_train = learner$predict(task, part$train)
p_test  = learner$predict(task, part$test)

surv = surv_breslow(times = task$times(part$train), status = task$status(part$train),
                    lp_train = p_train$lp, lp_test = p_test$lp)
head(surv)
}
\references{
Cox DR (1972).
\dQuote{Regression Models and Life-Tables.}
\emph{Journal of the Royal Statistical Society: Series B (Methodological)}, \bold{34}(2), 187--202.
\doi{10.1111/j.2517-6161.1972.tb00899.x}.

Lin, Y. D (2007).
\dQuote{On the Breslow estimator.}
\emph{Lifetime Data Analysis}, \bold{13}(4), 471-480.
\doi{10.1007/S10985-007-9048-Y/METRICS}.
}
