% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskDens.R
\name{TaskDens}
\alias{TaskDens}
\title{Density Task}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Task}.}
\description{
This task specializes \link{Task} for density estimation problems.
The target column is assumed to be numeric.
The \code{task_type} is set to \code{"density"}.

Predefined tasks are stored in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link{mlr_tasks}.
}
\section{Construction}{
\preformatted{t = TaskDens$new(id, backend, target)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the task.
\item \code{backend} :: (\link{DataBackend} | \code{data.frame()} | ...)\cr
Either a \link{DataBackend}, or any object which is convertible to a DataBackend with \code{as_data_backend()}.
E.g., a \code{data.frame()} will be converted to a \link{DataBackendDataTable}.
\item \code{target} :: \code{character(1)}\cr
Name of the target column.
}
}

\section{Fields}{

See \link{Task}.
}

\section{Methods}{

All methods from \link{Task}, and additionally:
\itemize{
\item \code{truth(rows = NULL)} :: \code{any}\cr
True response for specified \code{row_ids}. Format depends on the task type.
Defaults to all rows with role "use".
}
}

\examples{
task = TaskDens$new("precip", backend = data.frame(target = precip), target = "target")
task$task_type
task$truth()
}
\seealso{
Other Task: 
\code{\link{TaskSurv}}
}
\concept{Task}
