% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorSimdens.R
\name{TaskGeneratorSimdens}
\alias{TaskGeneratorSimdens}
\alias{mlr_task_generators_simdens}
\title{Density Task Generator for Package 'distr6'}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:TaskGenerator]{mlr3::TaskGenerator}.}
\description{
A \link[mlr3:TaskGenerator]{mlr3::TaskGenerator} calling \code{\link[distr6:distrSimulate]{distr6::distrSimulate()}} from package \CRANpkg{simsurv}.
See \code{\link[distr6:distrSimulate]{distr6::distrSimulate()}} for an explanation of the hyperparameters.
}
\section{Construction}{
\preformatted{TaskGeneratorSimdens$new()
mlr_task_generators$get("simdens")
tgen("simdens")
}
}

\examples{
generator = mlr3::mlr_task_generators$get("simdens")
task = generator$generate(20)
task$head()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:TaskGenerator]{TaskGenerators}: \link[mlr3:mlr_task_generators]{mlr3::mlr_task_generators}
}
