% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_scores.R
\name{weighted_survival_score}
\alias{weighted_survival_score}
\title{\itemize{
\item Either all of \code{times}, \code{t_max}, \code{p_max} are NULL, or only one of them is not
\item \code{times} is sorted (increasing), unique, positive time points
\item \code{t_max} > 0
\item \code{p_max} in \link{0,1}
\item \code{tmax_apply} = TRUE => one of \code{t_max}, \code{p_max} is given
\item \code{tmax_apply} = FALSE => \code{times} is given or all of \code{times}, \code{p_max} and \code{t_max} are NULL
The \code{t_max} cutoff will be applied later in the predicted survival matrix
to filter observations (rows) and time points (columns) + filter the
(time, status) target on both train (if provided) and test data
\strong{IMPORTANT}: times to calculate the score at => evaluation times
We start with the unique, sorted, test set time points
one of \code{t_max}, \code{p_max} is given
calculate \code{t_max} (time horizon) if \code{p_max} is given
check that \code{t_max} is within evaluation time range
\code{times} is given or it is \code{NULL}
is \code{times = NULL}, use the \code{unique_times}
}}
\usage{
weighted_survival_score(
  loss,
  truth,
  distribution,
  times = NULL,
  t_max = NULL,
  p_max = NULL,
  proper,
  train = NULL,
  eps,
  ...
)
}
\description{
\itemize{
\item Either all of \code{times}, \code{t_max}, \code{p_max} are NULL, or only one of them is not
\item \code{times} is sorted (increasing), unique, positive time points
\item \code{t_max} > 0
\item \code{p_max} in \link{0,1}
\item \code{tmax_apply} = TRUE => one of \code{t_max}, \code{p_max} is given
\item \code{tmax_apply} = FALSE => \code{times} is given or all of \code{times}, \code{p_max} and \code{t_max} are NULL
The \code{t_max} cutoff will be applied later in the predicted survival matrix
to filter observations (rows) and time points (columns) + filter the
(time, status) target on both train (if provided) and test data
\strong{IMPORTANT}: times to calculate the score at => evaluation times
We start with the unique, sorted, test set time points
one of \code{t_max}, \code{p_max} is given
calculate \code{t_max} (time horizon) if \code{p_max} is given
check that \code{t_max} is within evaluation time range
\code{times} is given or it is \code{NULL}
is \code{times = NULL}, use the \code{unique_times}
}
}
