% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_return.R
\name{.surv_return}
\alias{.surv_return}
\title{Get Survival Predict Types}
\usage{
.surv_return(
  times = NULL,
  surv = NULL,
  which.curve = NULL,
  crank = NULL,
  lp = NULL,
  response = NULL
)
}
\arguments{
\item{times}{(\code{numeric()}) \cr Vector of survival times.}

\item{surv}{(\code{matrix()|array()})\cr Matrix or array of predicted survival
probabilities, rows (1st dimension) are observations, columns (2nd dimension)
are times and in the case of an array there should be one more dimension.
Number of columns should be equal to length of \code{times}.}

\item{which.curve}{Which curve (3rd dimension) should the \code{crank} be
calculated for, in case \code{surv} is an \code{array}? If between (0,1) it is taken as
the quantile of the curves (assumes these are ordered) otherwise if greater
than 1 it is taken as the curve index. It can also be 'mean' and the survival
probabilities are averaged across the 3rd dimension.}

\item{crank}{(\code{numeric()})\cr Relative risk/continuous ranking. Higher value is associated
with higher risk. If \code{NULL} then either set as \code{-response} if available or
\code{lp} if available (this assumes that the \code{lp} prediction comes from a PH type
model - in case of an AFT model the user should provide \code{-lp}).
In case neither \code{response} or \code{lp} are provided, then \code{crank} is calculated
as the sum of the cumulative hazard function (expected mortality) derived from
the predicted survival function (\code{surv}). In case \code{surv} is a 3d array, we use
the \code{which.curve} parameter to decide which survival matrix (index in the 3rd
dimension) will be chosen for the calculation of \code{crank}.}

\item{lp}{(\code{numeric()})\cr Predicted linear predictor, used to impute \code{crank} if \code{NULL}.}

\item{response}{(\code{numeric()})\cr Predicted survival time, passed through function without
modification.}
}
\description{
Internal helper function to easily return the correct survival predict types.
}
\details{
Uses \link[survivalmodels:surv_to_risk]{survivalmodels::surv_to_risk} to reduce survival matrices to relative
risks / rankings if \code{crank} is NULL.
}
\references{
Sonabend, R., Bender, A., & Vollmer, S. (2022). Avoiding C-hacking when
evaluating survival distribution predictions with discrimination measures.
Bioinformatics. https://doi.org/10.1093/BIOINFORMATICS/BTAC451
}
