% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDE.R
\name{LearnerDensKDE}
\alias{LearnerDensKDE}
\alias{mlr_learners_dens.kde}
\title{Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls kernels implemented in \CRANpkg{distr6} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\details{
The default bandwidth uses Silverman's rule-of-thumb for Gaussian kernels, however for
non-Gaussian kernels it is recommended to use \CRANpkg{mlr3tuning} to tune the bandwidth with
cross-validation. Other density learners can be used for automated bandwidth selection.
}
\section{Construction}{
\preformatted{LearnerDensKDE$new()
mlr_learners$get("dens.kde")
lrn("dens.kde")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Packages: \CRANpkg{distr6}
}
}

\references{
\cite{mlr3proba}{silverman_1986}
}
\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensLogspline}},
\code{\link{LearnerDensMixed}},
\code{\link{LearnerDensPenalized}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
