% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensKDEgk.R
\name{LearnerDensKDEgk}
\alias{LearnerDensKDEgk}
\alias{mlr_learners_dens.kdeGK}
\title{Kernel Density Estimator}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerDens}.}
\description{
Calls \link[GenKern:KernSec]{GenKern::KernSec} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Construction}{
\preformatted{LearnerDensKDEgk$new()
mlr_learners$get("dens.kdeGK")
lrn("dens.kdeGK")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \code{pdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Properties: \code{missings}
\item Packages: \CRANpkg{GenKern} \CRANpkg{KernSmooth} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensLogspline}},
\code{\link{LearnerDensMixed}},
\code{\link{LearnerDensNonparametric}},
\code{\link{LearnerDensPenalized}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
