% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensPenalized.R
\name{LearnerDensPenalized}
\alias{LearnerDensPenalized}
\alias{mlr_learners_dens.pen}
\title{Penalized Density Estimator}
\description{
Calls \code{\link[pendensity:pendensity]{pendensity::pendensity()}} and the result is coerced to a \link[distr6:Distribution]{distr6::Distribution}.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link[mlr3:mlr_learners]{mlr_learners} or with the associated sugar function \link[mlr3:lrn]{lrn()}:\preformatted{LearnerDensPenalized$new()
mlr_learners$get("dens.pen")
lrn("dens.pen")
}
}

\section{Meta Information}{

\itemize{
\item Type: "dens"
\item Predict Types: \verb{pdf, cdf}
\item Feature Types: \verb{logical, integer, numeric, character, factor, ordered}
\item Properties: \code{-}
\item Packages: \CRANpkg{pendensity} \CRANpkg{distr6}
}
}

\seealso{
Other density estimators: 
\code{\link{LearnerDensHistogram}},
\code{\link{LearnerDensKDEkd}},
\code{\link{LearnerDensKDEks}},
\code{\link{LearnerDensKDE}},
\code{\link{LearnerDensLocfit}},
\code{\link{LearnerDensLogspline}},
\code{\link{LearnerDensMixed}},
\code{\link{LearnerDensNonparametric}},
\code{\link{LearnerDensPlugin}},
\code{\link{LearnerDensSpline}}
}
\concept{density estimators}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerDens]{mlr3proba::LearnerDens}} -> \code{LearnerDensPenalized}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerDensPenalized$new()}}
\item \href{#method-clone}{\code{LearnerDensPenalized$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensPenalized$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensPenalized$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
