% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvFlexible.R
\name{LearnerSurvFlexible}
\alias{LearnerSurvFlexible}
\alias{mlr_learners_surv.flexible}
\title{Flexible Parametric Spline Survival Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\description{
Flexible Parametric Spline Survival Learner
}
\details{
The \code{distr} return type is composed by using the formulae given from \code{\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline()}}. \cr
The \code{lp} return type is the fitted value of the location parameter, \code{gamma0}. \cr
The \code{crank} return type is the same as the \code{lp}. \cr

The predict method is based on \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}} but is quicker and more
efficient, and adapts the return type to be compatible with \CRANpkg{distr6}.
}
\section{Construction}{
\preformatted{LearnerSurvFlexible$new()
mlr_learners$get("surv.flexible")
lrn("surv.flexible")
}
}

\references{
Royston, P. and Parmar, M. (2002).
Flexible parametric proportional-hazards and proportional-odds models for censored survival data,
with application to prognostic modelling and estimation of treatment effects.
Statistics in Medicine, 21(15), 2175-2197.
\doi{10.1002/sim.1203}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
