% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvFlexible.R
\name{LearnerSurvFlexible}
\alias{LearnerSurvFlexible}
\alias{mlr_learners_surv.flexible}
\title{Flexible Parametric Spline Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Calls \code{\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline()}}.
\itemize{
\item lp is predicted by using an internally defined \code{predict} method, see details
\item distr is predicted by using an internally defined \code{predict} method, see details
\item \code{crank} is identical to \code{lp}
}

Parameter \code{k} is changed to \code{1} and \code{scale} is changed to \code{odds}, as these are more in line with
the Royston/Parmar proposed models, and the package defaults are equivalent to fitting a
parametric model and therefore \link[=LearnerSurvParametric]{surv.parametric} should be used instead.

If fitting a model with \code{k = 0} then consider using \link[=LearnerSurvParametric]{surv.parametric} as
this is likely to have more optimal results, and has more options for tuning.
}
\details{
The \code{distr} prediction is estimated using the fitted custom distributions
from \code{\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline()}} and the estimated coefficients.

As flexible spline models estimate the baseline hazard as the intercept, the linear predictor,
\code{lp}, can be calculated as in the classical setting. i.e. For fitted coefficients,
\eqn{\beta = (\beta_0,...,\beta_P)}{\beta = (\beta0,...,\betaP)},
and covariates \eqn{X^T = (X_0,...,X_P)^T}{X^T = (X0,...,XP)^T}, where \eqn{X_0}{X0} is a column
of \eqn{1}s: \eqn{lp = \beta X}{lp = \betaX}.
}
\section{Construction}{
\preformatted{LearnerSurvFlexible$new()
mlr_learners$get("surv.flexible")
lrn("surv.flexible")
}
}

\section{Meta Information}{

\itemize{
\item Type: "surv"
\item Predict Types: \verb{distr, lp, crank}
\item Feature Types: \verb{logical, integer, factor, numeric}
\item Packages: \CRANpkg{flexsurv} \CRANpkg{survival} \CRANpkg{distr6}
}
}

\references{
Royston, P. and Parmar, M. (2002).
Flexible parametric proportional-hazards and proportional-odds models for censored survival data,
with application to prognostic modelling and estimation of treatment effects.
Statistics in Medicine, 21(15), 2175-2197.
\doi{10.1002/sim.1203}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}

Other survival learners: 
\code{\link{LearnerSurvBlackboost}},
\code{\link{LearnerSurvCVGlmnet}},
\code{\link{LearnerSurvCoxPH}},
\code{\link{LearnerSurvGBM}},
\code{\link{LearnerSurvGamboost}},
\code{\link{LearnerSurvGlmboost}},
\code{\link{LearnerSurvGlmnet}},
\code{\link{LearnerSurvKaplan}},
\code{\link{LearnerSurvMboost}},
\code{\link{LearnerSurvNelson}},
\code{\link{LearnerSurvParametric}},
\code{\link{LearnerSurvPenalized}},
\code{\link{LearnerSurvRandomForestSRC}},
\code{\link{LearnerSurvRanger}},
\code{\link{LearnerSurvRpart}},
\code{\link{LearnerSurvSVM}}
}
\concept{survival learners}
