% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvParametric.R
\name{LearnerSurvParametric}
\alias{LearnerSurvParametric}
\alias{mlr_learners_surv.parametric}
\title{Fully Parametric Survival Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerSurv}.}
\description{
A \link{LearnerSurv} for a Fully Parametric model partially implemented in
\code{\link[survival:survreg]{survival::survreg()}} in package \CRANpkg{survival}.
}
\details{
The \code{distr} return type is composed by using the formulae for proportional hazard (PH),
accelerated failure time (AFT), or proportional odds (PO) models, with the choice dependent on the
\code{type} hyper-parameter. \cr
The \code{crank} return type is defined as the exponential of the linear predictor. This ranking
is identical to the expectation of the survival distribution but faster and more accurate to compute. \cr
The \code{lp} return type is predicted by computing \eqn{X\beta} for the coefficients \eqn{\beta}
fit in \code{\link[survival:survreg]{survival::survreg()}}.

The predict method is based on \code{\link[survival:predict.survreg]{survival::predict.survreg()}} but is more efficient for composition
to distributions.

Currently six parameterisations can be assumed for the baseline, see \code{\link[survival:survreg]{survival::survreg()}}. These
are internally re-parameterised and defined as \code{distr6} objects.
}
\section{Construction}{
\preformatted{LearnerSurvParametric$new()
mlr_learners$get("mlr_learners_surv.parametric")
lrn("mlr_learners_surv.parametric")
}
}

\references{
Kalbfleisch, J. D., Prentice, R. L. (2002).
The Statistical Analysis of Failure Time Data.
John Wiley & Sons.
\doi{10.1002/9781118032985}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
