% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvPenalized.R
\docType{data}
\name{LearnerSurvPenalized}
\alias{LearnerSurvPenalized}
\alias{mlr_learners_surv.penalized}
\title{L1 and L2 Penalized Estiamtion in GLMs Survival Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
Generalized linear models with elastic net regularization.
Calls \code{\link[penalized:penalized]{penalized::penalized()}} from package \CRANpkg{penalized}.
}
\section{Construction}{
\preformatted{LearnerSurvPenalized$new()
mlr_learners$get("surv.penalized")
lrn("surv.penalized")
}
}

\examples{
library(mlr3)
task = tgen("simsurv")$generate(200)
learner = lrn("surv.penalized")
resampling = rsmp("cv", folds = 3)
resample(task, learner, resampling)
}
\references{
Goeman, J. J., L1 penalized estimation in the Cox proportional hazards model.
Biometrical Journal 52(1), 70{84}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
