% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSurvRandomForestSRC.R
\docType{data}
\name{LearnerSurvRandomForestSRC}
\alias{LearnerSurvRandomForestSRC}
\alias{mlr_learners_surv.randomForestSRC}
\title{Survival RandomForestSRC Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerSurv}.}
\description{
A \link{LearnerSurv} for a survival random forest implemented in \code{\link[randomForestSRC:rfsrc]{randomForestSRC::rfsrc()}} in package \CRANpkg{randomForestSRC}.
}
\details{
\code{\link[randomForestSRC]{rfsrc}} has three prediction outcomes, from the fitted model these are
respectively:
\enumerate{
\item predicted - This is ensemble mortality defined in Ishwaran et al. (2008), the sum
of an individuals cumulative hazard function over all time-points
\item chf - Cumulative hazard function, estimated via a bootstrapped Nelson-Aalen estimator (Ishwaran, 2008)
\item surv - Survival function, estimated via a bootrstrapped Kaplan-Meier estimate (https://kogalur.github.io/randomForestSRC/theory.html)
}

Only the second two of these are returned in the 'distr' predict.type, as Nelson-Aalen and Kaplan-Meier
will give different results, the estimator can be chosen in the parameter set under "estimator".

The 'risk' predict.type is here defined as the mean of the cumulative hazard over all unique death times.
}
\section{Construction}{
\preformatted{LearnerSurvRandomForestSRC$new()
mlr_learners$get("surv.randomForestSRC")
lrn("surv.randomForestSRC")
}
}

\references{
Ishwaran H. and Kogalur U.B. (2019). Fast Unified Random Forests for Survival,
Regression, and Classification (RF-SRC), R package version 2.9.1.

Ishwaran H., Kogalur U.B., Blackstone E.H. and Lauer M.S. (2008). Random survival
forests. Ann. Appl. Statist. 2(3), 841--860.

Breiman, L. (2001).
Random Forests.
Machine Learning 45(1).
\doi{10.1023/A:1010933404324}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
