% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureDens.R
\name{MeasureDens}
\alias{MeasureDens}
\title{Density Measure}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Measure]{mlr3::Measure}.}
\description{
This measure specializes \link[mlr3:Measure]{mlr3::Measure} for density estimation problems.
Predefined measures can be found in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link[mlr3:mlr_measures]{mlr3::mlr_measures}.
}
\section{Construction}{
\preformatted{MeasureDens$new(id, range, minimize = NA, aggregator = NULL, properties = character(),
predict_type = "pdf", task_properties = character(0L), packages = character(0L))
}

For a description of the arguments, see \link[mlr3:Measure]{mlr3::Measure}.
The \code{task_type} is set to \code{"dens"}.
Possible values for \code{predict_type} are \code{"pdf"} and \code{"cdf"}.
}

\section{Fields}{

See \link{Measure}.
}

\section{Methods}{

See \link{Measure}.
}

\seealso{
Other Measure: 
\code{\link{MeasureSurv}}
}
\concept{Measure}
