% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureDensLogloss.R
\name{MeasureDensLogloss}
\alias{MeasureDensLogloss}
\alias{mlr_measures_dens.logloss}
\title{Log loss Density Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureDens}.}
\description{


Calculates the cross-entropy, or logarithmic (log), loss.

The logloss, in the context of probabilistic predictions, is defined as the negative log probability
density function, \eqn{f}, evaluated at the observed value, \eqn{y},
\deqn{L(f, y) = -log(f(y))}
}
\section{Construction}{
\preformatted{MeasureDensLogloss$new(eps = 1e-15)
mlr_measures$get("dens.logloss")
msr("dens.logloss")
}
\itemize{
\item \code{eps} :: \code{numeric(1)} \cr Very small number to set zero-valued predicted probabilities to, in order to prevent errors in log(0) calculation. \cr
}
}

\section{Meta Information}{

\itemize{
\item Type: \code{"density"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{pdf}
}
}

\section{Fields}{

See \link{MeasureDens}, as well as all variables passed to the constructor.


As well as
\itemize{
\item eps :: numeric(1) \cr
Very small number to set zero-valued predicted probabilities to, in order to prevent errors in log(0) calculation.
}
}

\concept{Density estimation measures}
\concept{density measures}
