% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvGraf.R
\name{MeasureSurvGraf}
\alias{MeasureSurvGraf}
\alias{mlr_measures_surv.graf}
\alias{MeasureSurvBrier}
\alias{mlr_measures_surv.brier}
\title{Integrated Graf Score Survival Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{


Calculates the Integrated Graf Score (IGS), aka integrated Brier survival score or squared loss.
An approximation to the IGS is calculated by taking the mean over all time-points in the test set.

For an individual who dies at time \eqn{t}, with predicted Survival function, \eqn{S}, the
(unweighted) Graf Score at time t* is given by
\deqn{G(S, t*) = (I(t > t*) - S(t*))^2}

To account for censoring a weighted version is defined by
\deqn{G(S, t*) = S(t*)^2 * I(t \le t*, \delta = 1) * (1/G(t)) + (1 - S(t*))^2 * I(t > t*) * (1/G(t*))}

As only a finite number of test points are given, an approximation to the IGS can be made by taking
the average over all, \eqn{M}, unique discrete time-points given in the test data,
\deqn{IGS(S) = 1/M \Sigma_t G(S, t)}

Finally the sample mean is taken to return a single score for all \eqn{N} observations,
\deqn{IGS(pred) = 1/N \Sigma_i IGS(S_i)}
}
\section{Construction}{
\preformatted{MeasureSurvGraf$new(integrated = TRUE, times)
mlr_measures$get("surv.graf")
msr("surv.graf")
}
\itemize{
\item \code{integrated} :: \code{logical(1)} \cr If \code{TRUE} (default), returns the integrated score, otherwise returns the score corresponding to the \code{times} argument. \cr
\item \code{times} :: \code{vector()} \cr Numeric vector of times where to evaluate the score (provide only one for best performance). If missing then returns score at all unique time-points in test set. \cr
}
}

\section{Meta Information}{

\itemize{
\item Type: \code{"surv"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{distr}
}
}

\references{
Graf, E., Schmoor, C., Sauerbrei, W. and Schumacher, M. (1999).\cr
Assessment and comparison of prognostic classification schemes for survival data.\cr
Statistics in Medicine, 18(17), 2529-2545.\cr
\doi{10.1002/(SICI)1097-0258(19990915/30)18:17/18<2529::AID-SIM274>3.0.CO;2-5}
}
\seealso{
Other survival measures: 
\code{\link{MeasureSurvBeggC}},
\code{\link{MeasureSurvChamblessAUC}},
\code{\link{MeasureSurvGonenC}},
\code{\link{MeasureSurvGrafSE}},
\code{\link{MeasureSurvHarrellC}},
\code{\link{MeasureSurvHungAUC}},
\code{\link{MeasureSurvLoglossIntSE}},
\code{\link{MeasureSurvLoglossInt}},
\code{\link{MeasureSurvLoglossSE}},
\code{\link{MeasureSurvLogloss}},
\code{\link{MeasureSurvNagelkR2}},
\code{\link{MeasureSurvOQuigleyR2}},
\code{\link{MeasureSurvSongAUC}},
\code{\link{MeasureSurvSongTNR}},
\code{\link{MeasureSurvSongTPR}},
\code{\link{MeasureSurvUnoAUC}},
\code{\link{MeasureSurvUnoC}},
\code{\link{MeasureSurvUnoTNR}},
\code{\link{MeasureSurvUnoTPR}},
\code{\link{MeasureSurvXuR2}}

Other Probabilistic survival measures: 
\code{\link{MeasureSurvGrafSE}},
\code{\link{MeasureSurvLoglossIntSE}},
\code{\link{MeasureSurvLoglossInt}},
\code{\link{MeasureSurvLoglossSE}},
\code{\link{MeasureSurvLogloss}}
}
\concept{Probabilistic survival measures}
\concept{survival measures}
