% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvGrafSE.R
\name{MeasureSurvGrafSE}
\alias{MeasureSurvGrafSE}
\alias{mlr_measures_surv.grafSE}
\title{Graf Score}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calculates the standard error of \link{MeasureSurvGraf}.
}
\section{Construction}{
\preformatted{MeasureSurvGrafSE$new()
mlr_measures$get("surv.grafSE")
msr("surv.grafSE")
}
}

\references{
Graf, G. (1950).
Verification of forecasts expressed in terms of probability.
Monthly Weather Review, 78(1), 1-3.
\doi{10.1175/1520-0493(1950)078<0001:VOFEIT>2.0.CO;2}

Graf, E., Schmoor, C., Sauerbrei, W. and Schumacher, M. (1999).
Assessment and comparison of prognostic classification schemes for survival data.
Statistics in Medicine, 18(17), 2529-2545.
\doi{10.1002/(SICI)1097-0258(19990915/30)18:17/18<2529::AID-SIM274>3.0.CO;2-5}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
