% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvIntLoglossSE.R
\name{MeasureSurvLoglossIntSE}
\alias{MeasureSurvLoglossIntSE}
\alias{mlr_measures_surv.loglossintSE}
\title{Standard Error of Integrated Logloss}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calculates the standard error of \link{MeasureSurvLoglossInt}.
}
\section{Construction}{
\preformatted{MeasureSurvLoglossIntSE$new()
mlr_measures$get("surv.loglossintSE")
msr("surv.loglossintSE")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
