% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvLoglossSE.R
\name{MeasureSurvLoglossSE}
\alias{MeasureSurvLoglossSE}
\alias{mlr_measures_surv.logloss_se_se}
\title{Standard Error of Logloss}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calculates the standard error of the cross-entropy, or logarithmic, loss.
}
\section{Construction}{
\preformatted{MeasureSurvLoglossSESE$new()
mlr_measures$get("surv.logloss_se")
msr("surv.logloss_se")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
