% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvNagelkR2.R
\name{MeasureSurvNagelkR2}
\alias{MeasureSurvNagelkR2}
\alias{mlr_measures_surv.nagelkR2}
\title{Nagelkerke's R2 Coefficient}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:Nagelk]{survAUC::Nagelk()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes Cox PH model specification.
}
\section{Construction}{
\preformatted{MeasureSurvNagelkR2$new()
mlr_measures$get("surv.nagelkR2")
msr("surv.nagelkR2")
}
}

\references{
Nagelkerke, N. J. D. (1991).
A note on a general definition of the coefficient of determination.
Biometrika 78, 691–692.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
