% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvUnoTNR.R
\name{MeasureSurvUnoTNR}
\alias{MeasureSurvUnoTNR}
\alias{mlr_measures_surv.unoTNR}
\title{Uno's TNR}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:spec.uno]{survAUC::spec.uno()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes random censoring.
}
\section{Construction}{
\preformatted{MeasureSurvUnoTNR$new()
mlr_measures$get("surv.unoTNR")
msr("surv.unoTNR")
}
}

\references{
Uno, H., T. Cai, L. Tian, and L. J. Wei (2007).
Evaluating prediction rules for t-year survivors with censored regression models.
Journal of the American Statistical Association 102, 527–537.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
