% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvUnoTPR.R
\name{MeasureSurvUnoTPR}
\alias{MeasureSurvUnoTPR}
\alias{mlr_measures_surv.unoTPR}
\title{Uno's TPR}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureSurv}.}
\description{
Calls \code{\link[survAUC:sens.uno]{survAUC::sens.uno()}}.
}
\details{
Requires \code{lp} \code{predict_type}. \cr
Assumes random censoring.
}
\section{Construction}{
\preformatted{MeasureSurvUnoTPR$new()
mlr_measures$get("surv.unoTPR")
msr("surv.unoTPR")
}
}

\references{
Uno, H., T. Cai, L. Tian, and L. J. Wei (2007).
Evaluating prediction rules for t-year survivors with censored regression models.
Journal of the American Statistical Association 102, 527–537.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Measure]{Measures}: \link[mlr3:mlr_measures]{mlr3::mlr_measures}
}
