% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3proba-package}
\alias{mlr3proba}
\alias{mlr3proba-package}
\title{mlr3proba: Probabilistic Supervised Learning}
\description{
Placeholder description.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk} (0000-0001-9225-4654)

Authors:
\itemize{
  \item Franz Kiraly \email{f.kiraly@ucl.ac.uk}
}

}
