% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3proba-package}
\alias{mlr3proba}
\alias{mlr3proba-package}
\title{mlr3proba: Probabilistic Supervised Learning for 'mlr3'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides extensions for probabilistic supervised
    learning for 'mlr3'.  This includes extending the regression task to
    probabilistic and interval regression, adding a survival task, and
    other specialized models, predictions, and measures. mlr3extralearners is available
    from <https://github.com/mlr-org/mlr3extralearners>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3proba.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3proba}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3proba/issues}
}

}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk} (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})

Authors:
\itemize{
  \item Franz Kiraly \email{f.kiraly@ucl.ac.uk}
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
}

Other contributors:
\itemize{
  \item Nurul Ain Toha \email{nurul.toha.15@ucl.ac.uk} [contributor]
  \item Andreas Bender \email{bender.at.R@gmail.com} (\href{https://orcid.org/0000-0001-5628-8611}{ORCID}) [contributor]
}

}
