% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSurvGraf.R
\name{mlr_measures_surv.graf}
\alias{mlr_measures_surv.graf}
\alias{MeasureSurvGraf}
\alias{MeasureSurvBrier}
\alias{mlr_measures_surv.brier}
\title{Integrated Brier Score Survival Measure}
\description{
Calculates the \strong{Integrated Survival Brier Score} (ISBS), Integrated Graf Score
or squared survival loss.
}
\details{
For an individual who dies at time \eqn{t}, with predicted Survival function, \eqn{S}, the
Graf Score at time \eqn{t^*}{t*} is given by
\deqn{L_{ISBS}(S,t|t^*) = [(S(t^*)^2)I(t \le t^*, \delta = 1)(1/G(t))] + [((1 - S(t^*))^2)I(t > t^*)(1/G(t^*))]}
where \eqn{G} is the Kaplan-Meier estimate of the censoring distribution.

The re-weighted ISBS (RISBS) is given by
\deqn{L_{RISBS}(S,t|t^*) = [(S(t^*)^2)I(t \le t^*, \delta = 1)(1/G(t))] + [((1 - S(t^*))^2)I(t > t^*)(1/G(t))]}
where \eqn{G} is the Kaplan-Meier estimate of the censoring distribution, i.e. always
weighted by \eqn{G(t)}.
RISBS is strictly proper when the censoring distribution is independent
of the survival distribution and when G is fit on a sufficiently large dataset.
ISBS is never proper. Use \code{proper = FALSE} for ISBS and \code{proper = TRUE} for RISBS.
Results may be very different if many observations are
censored at the last observed time due to division by 1/\code{eps} in \code{proper = TRUE}.

\strong{Note}: If comparing the integrated graf score to other packages, e.g.
\CRANpkg{pec}, then \code{method = 2} should be used. However the results may
still be very slightly different as this package uses \code{survfit} to estimate
the censoring distribution, in line with the Graf 1999 paper; whereas some
other packages use \code{prodlim} with \code{reverse = TRUE} (meaning Kaplan-Meier is
not used).

If \code{task} and \code{train_set} are passed to \verb{$score} then G is fit on training data,
otherwise testing data. The first is likely to reduce any bias caused by calculating
parts of the measure on the test data it is evaluating. The training data is automatically
used in scoring resamplings.
}
\section{Dictionary}{

This \link[mlr3:Measure]{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link[mlr3:mlr_measures]{mlr_measures} or with the associated sugar function \link[mlr3:mlr_sugar]{msr()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{MeasureSurvGraf$new()
mlr_measures$get("surv.graf")
msr("surv.graf")
}\if{html}{\out{</div>}}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   integrated \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   times \tab untyped \tab - \tab  \tab - \cr
   t_max \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   p_max \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   method \tab integer \tab 2 \tab  \tab \eqn{[1, 2]}{[1, 2]} \cr
   se \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   proper \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   eps \tab numeric \tab 0.001 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   ERV \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
}
}

\section{Meta Information}{

\itemize{
\item Type: \code{"surv"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{distr}
}
}

\section{Parameter details}{

\itemize{
\item \code{integrated} (\code{logical(1)})\cr
If \code{TRUE} (default), returns the integrated score (eg across
time points); otherwise, not integrated (eg at a single time point).
}


\itemize{
\item \code{times} (\code{numeric()})\cr
If \code{integrate == TRUE} then a vector of time-points over which to integrate the score.
If \code{integrate == FALSE} then a single time point at which to return the score.
}


\itemize{
\item \code{t_max} (\code{numeric(1)})\cr
Cutoff time (i.e. time horizon) to evaluate the measure up to.
Mutually exclusive with \code{p_max} or \code{times}.
}


\itemize{
\item \code{p_max} (\code{numeric(1)})\cr
The proportion of censoring to integrate up to in the given dataset.
Mutually exclusive with \code{times} or \code{t_max}.
}


\itemize{
\item \code{method} (\code{integer(1)})\cr
If \code{integrate == TRUE}, this selects the integration weighting method.
\code{method == 1} corresponds to weighting each time-point equally
and taking the mean score over discrete time-points.
\code{method == 2} corresponds to calculating a mean weighted by the
difference between time-points.
\code{method == 2} is the default value, to be in line with other packages.
}


\itemize{
\item \code{se} (\code{logical(1)})\cr
If \code{TRUE} then returns standard error of the measure otherwise
returns the mean across all individual scores, e.g. the mean of
the per observation scores.
Default is \code{FALSE} (returns the mean).
}


\itemize{
\item \code{proper} (\code{logical(1)})\cr
If \code{TRUE} then weights scores by the censoring distribution at
the observed event time, which results in a strictly proper scoring
rule if censoring and survival time distributions are independent
and a sufficiently large dataset is used.
If \code{FALSE} then weights scores by the Graf method which is the
more common usage but the loss is not proper.
}


\itemize{
\item \code{eps} (\code{numeric(1)})\cr
Very small number to substitute zero values in order to prevent errors
in e.g. log(0) and/or division-by-zero calculations.
Default value is 0.001.
}


\itemize{
\item \code{ERV} (\code{logical(1)})\cr
If \code{TRUE} then the Explained Residual Variation method is applied, which
means the score is standardized against a Kaplan-Meier baseline.
Default is \code{FALSE}.
}
}

\references{
Graf E, Schmoor C, Sauerbrei W, Schumacher M (1999).
\dQuote{Assessment and comparison of prognostic classification schemes for survival data.}
\emph{Statistics in Medicine}, \bold{18}(17-18), 2529--2545.
\doi{10.1002/(sici)1097-0258(19990915/30)18:17/18<2529::aid-sim274>3.0.co;2-5}.
}
\seealso{
Other survival measures: 
\code{\link{mlr_measures_surv.calib_alpha}},
\code{\link{mlr_measures_surv.calib_beta}},
\code{\link{mlr_measures_surv.chambless_auc}},
\code{\link{mlr_measures_surv.cindex}},
\code{\link{mlr_measures_surv.dcalib}},
\code{\link{mlr_measures_surv.hung_auc}},
\code{\link{mlr_measures_surv.intlogloss}},
\code{\link{mlr_measures_surv.logloss}},
\code{\link{mlr_measures_surv.mae}},
\code{\link{mlr_measures_surv.mse}},
\code{\link{mlr_measures_surv.nagelk_r2}},
\code{\link{mlr_measures_surv.oquigley_r2}},
\code{\link{mlr_measures_surv.rcll}},
\code{\link{mlr_measures_surv.rmse}},
\code{\link{mlr_measures_surv.schmid}},
\code{\link{mlr_measures_surv.song_auc}},
\code{\link{mlr_measures_surv.song_tnr}},
\code{\link{mlr_measures_surv.song_tpr}},
\code{\link{mlr_measures_surv.uno_auc}},
\code{\link{mlr_measures_surv.uno_tnr}},
\code{\link{mlr_measures_surv.uno_tpr}},
\code{\link{mlr_measures_surv.xu_r2}}

Other Probabilistic survival measures: 
\code{\link{mlr_measures_surv.intlogloss}},
\code{\link{mlr_measures_surv.logloss}},
\code{\link{mlr_measures_surv.rcll}},
\code{\link{mlr_measures_surv.schmid}}

Other distr survival measures: 
\code{\link{mlr_measures_surv.calib_alpha}},
\code{\link{mlr_measures_surv.dcalib}},
\code{\link{mlr_measures_surv.intlogloss}},
\code{\link{mlr_measures_surv.logloss}},
\code{\link{mlr_measures_surv.rcll}},
\code{\link{mlr_measures_surv.schmid}}
}
\concept{Probabilistic survival measures}
\concept{distr survival measures}
\concept{survival measures}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3proba:MeasureSurv]{mlr3proba::MeasureSurv}} -> \code{MeasureSurvGraf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureSurvGraf-new}{\code{MeasureSurvGraf$new()}}
\item \href{#method-MeasureSurvGraf-clone}{\code{MeasureSurvGraf$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3proba" data-topic="MeasureSurv" data-id="print"><a href='../../mlr3proba/html/MeasureSurv.html#method-MeasureSurv-print'><code>mlr3proba::MeasureSurv$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureSurvGraf-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureSurvGraf-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureSurvGraf$new(ERV = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ERV}}{(\code{logical(1)})\cr
Standardize measure against a Kaplan-Meier baseline
(Explained Residual Variation)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureSurvGraf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureSurvGraf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureSurvGraf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
