% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpTaskSurvClassifDiscTime.R
\name{mlr_pipeops_trafotask_survclassif_disctime}
\alias{mlr_pipeops_trafotask_survclassif_disctime}
\alias{PipeOpTaskSurvClassifDiscTime}
\title{PipeOpTaskSurvClassifDiscTime}
\description{
Transform \link{TaskSurv} to \link[mlr3:TaskClassif]{TaskClassif} by dividing continuous
time into multiple time intervals for each observation.
This transformation creates a new target variable \code{ped_status} that indicates
whether an event occurred within each time interval.
This approach facilitates survival analysis within a classification framework
using discrete time intervals (Tutz et al. 2016).
}
\section{Input and Output Channels}{

\link{PipeOpTaskSurvClassifDiscTime} has one input channel named "input", and two
output channels, one named "output" and the other "transformed_data".

During training, the "output" is the "input" \link{TaskSurv} transformed to a
\link[mlr3:TaskClassif]{TaskClassif}.
The target column is named \code{ped_status} and indicates whether an event occurred
in each time interval.
An additional feature named \code{tend} is added to the ouput task, containing the
end time of each interval.
The "transformed_data" is an empty \link[data.table:data.table]{data.table}.

During prediction, the "input" \link{TaskSurv} is transformed to the "output"
\link[mlr3:TaskClassif]{TaskClassif} with \code{ped_status} as target and the \code{tend}
feature included.
The "transformed_data" is a \link{data.table} which has all the features of the
"output" task, including an additional column \code{time2} containing the
original times.
This "transformed_data" is only meant to be used with the \link{PipeOpPredClassifSurvDiscTime}.
}

\section{State}{

The \verb{$state} contains information about the \code{cut} parameter used
as well as \code{time_var} and \code{event_var}, the names of the two target
columns of the survival task.
}

\section{Parameters}{

The parameters are
\itemize{
\item \code{cut :: numeric()}\cr
Split points, used to partition the data into intervals based on the \code{time} column.
If unspecified, all unique event times will be used.
If \code{cut} is a single integer, it will be interpreted as the number of equidistant
intervals from 0 until the maximum event time.
\item \code{max_time :: numeric(1)}\cr
If \code{cut} is unspecified, this will be the last possible event time.
All event times after \code{max_time} will be administratively censored at \code{max_time.}
Needs to be greater than the minimum event time in the given task.
}
}

\examples{
\dontrun{
if (requireNamespace("mlr3pipelines", quietly = TRUE) &&
    requireNamespace("mlr3learners", quietly = TRUE)) {
  library(mlr3)
  library(mlr3learners)
  library(mlr3pipelines)

  task = tsk("lung")

  # transform the survival task to a classification task
  po_disc = po("trafotask_survclassif_disctime", cut = 4)
  task_classif = po_disc$train(list(task))[[1L]]

  # use a classification learner
  learner = lrn("classif.log_reg", predict_type = "prob")
  learner$train(task_classif)
  learner$predict(task_classif)
}
}

}
\references{
Tutz, Gerhard, Schmid, Matthias (2016).
\emph{Modeling Discrete Time-to-Event Data},  series Springer Series in Statistics.
Springer International Publishing.
ISBN 978-3-319-28156-8 978-3-319-28158-2, \url{http://link.springer.com/10.1007/978-3-319-28158-2}.
}
\seealso{
Other PipeOps: 
\code{\link{PipeOpPredTransformer}},
\code{\link{PipeOpTaskTransformer}},
\code{\link{PipeOpTransformer}},
\code{\link{mlr_pipeops_survavg}},
\code{\link{mlr_pipeops_trafopred_classifsurv_disctime}},
\code{\link{mlr_pipeops_trafopred_regrsurv}},
\code{\link{mlr_pipeops_trafopred_survregr}},
\code{\link{mlr_pipeops_trafotask_regrsurv}},
\code{\link{mlr_pipeops_trafotask_survregr}}

Other Transformation PipeOps: 
\code{\link{mlr_pipeops_trafopred_classifsurv_disctime}},
\code{\link{mlr_pipeops_trafopred_regrsurv}},
\code{\link{mlr_pipeops_trafopred_survregr}},
\code{\link{mlr_pipeops_trafotask_regrsurv}},
\code{\link{mlr_pipeops_trafotask_survregr}}
}
\concept{PipeOps}
\concept{Transformation PipeOps}
\section{Super class}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{PipeOpTaskSurvClassifDiscTime}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpTaskSurvClassifDiscTime-new}{\code{PipeOpTaskSurvClassifDiscTime$new()}}
\item \href{#method-PipeOpTaskSurvClassifDiscTime-clone}{\code{PipeOpTaskSurvClassifDiscTime$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpTaskSurvClassifDiscTime-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpTaskSurvClassifDiscTime-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpTaskSurvClassifDiscTime$new(id = "trafotask_survclassif_disctime")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier of the resulting  object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpTaskSurvClassifDiscTime-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpTaskSurvClassifDiscTime-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpTaskSurvClassifDiscTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
