% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv_X.R
\name{mlr_tasks_unemployment}
\alias{mlr_tasks_unemployment}
\title{Unemployment Duration Survival Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.
}
\description{
A survival task for the \code{UnempDur} data set.
}
\details{
A survival task for the "UnempDur" data set in package \CRANpkg{Ecdat}.
Contains the following columns of the original data set:
"spell" (time), "censor1" (status), "age", "ui", "logwage", and "tenure".
}
\section{Construction}{
\preformatted{mlr3::mlr_tasks$get("unemployment")
mlr3::tsk("unemployment")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Task]{Tasks}: \link[mlr3:mlr_tasks]{mlr3::mlr_tasks}
}
