% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_zzz.R
\name{mlr_tasks_whas}
\alias{mlr_tasks_whas}
\title{Worcester Heart Attack Study (WHAS) Survival Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskSurv}.
}
\description{
A survival task for the \link{whas} data set.
}
\section{Dictionary}{

This \link{Task} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_tasks} or with the associated sugar function \code{\link[=tsk]{tsk()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_tasks$get("whas")
tsk("whas")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Dimensions: 481x11
\item Properties: -
\item Has Missings: \code{FALSE}
\item Target: \dQuote{time}, \dQuote{status}
\item Features: \dQuote{age}, \dQuote{chf}, \dQuote{cpk}, \dQuote{lenstay}, \dQuote{miord}, \dQuote{mitype}, \dQuote{sexF}, \dQuote{sho}, \dQuote{year}
}
}

\section{Preprocessing}{

\itemize{
\item Columns \code{id}, \code{yrgrp}, and \code{dstat} are removed.
\item Column \code{sex} is renamed to \code{sexF}, \code{lenfol} to \code{time}, and \code{fstat} to \code{status}.
\item Target is total follow-up time from hospital admission.
}
}

\seealso{
\itemize{
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}:
\url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html}
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Task]{Tasks}: \link[mlr3:mlr_tasks]{mlr3::mlr_tasks}
\item \code{as.data.table(mlr_tasks)} for a table of available \link[=Task]{Tasks} in the running session
}

Other Task: 
\code{\link{TaskCompRisks}},
\code{\link{TaskDens}},
\code{\link{TaskSurv}},
\code{\link{mlr_tasks_actg}},
\code{\link{mlr_tasks_faithful}},
\code{\link{mlr_tasks_gbcs}},
\code{\link{mlr_tasks_gbsg}},
\code{\link{mlr_tasks_grace}},
\code{\link{mlr_tasks_lung}},
\code{\link{mlr_tasks_mgus}},
\code{\link{mlr_tasks_pbc}},
\code{\link{mlr_tasks_precip}},
\code{\link{mlr_tasks_rats}},
\code{\link{mlr_tasks_veteran}}
}
\concept{Task}
