% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionSurv.R
\name{PredictionSurv}
\alias{PredictionSurv}
\title{Prediction Object for Survival}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3:Prediction]{mlr3::Prediction}.}
\description{
This object stores the predictions returned by a learner of class \link{LearnerSurv}.

The \code{task_type} is set to \code{"surv"}.
}
\section{Construction}{
\preformatted{p = PredictionSurv$new(task = NULL, row_ids = task$row_ids, truth = task$truth(), distr = NULL,
risk = NULL, lp = NULL)
}
\itemize{
\item \code{task} :: \link{TaskSurv}\cr
Task, used to extract defaults for \code{row_ids} and \code{truth}.
\item \code{row_ids} :: (\code{integer()} | \code{character()})\cr
Row ids of the task. Per default, these are extracted from the \code{task}.
\item \code{truth} :: \code{survival::Surv()}\cr
Observed survival times. Per default, these are extracted from the \code{task}.
\item \code{risk} :: \code{numeric()}\cr
Vector of risk scores. One element for each observation in the test set.
The higher the risk, the more likely is an event.
Used in measures like \link{mlr_measures_surv.harrells_c}.
\item \code{lp} :: \code{numeric()}\cr
Vector of linear predictor scores. One element for each observation in the test set.
lp = exp(risk).
Used in measures like \link{mlr_measures_surv.harrells_c}.
\item \code{distr} :: \code{distr6::Distribution()}\cr
List of R6 distributions. One distribution for each observation in the test set.
Each distribution contains the hazard, survival, and cumulative hazard (and other
common functions) for all predictions.
Used in measures like \link{mlr_measures_surv.brier}.
}
}

\section{Fields}{

See \link[mlr3:Prediction]{mlr3::Prediction}.

The field \code{task_type} is set to \code{"surv"}.
}

\examples{
library(mlr3)
task = mlr_tasks$get("lung")
learner = mlr_learners$get("surv.rpart")
p = learner$train(task)$predict(task)
head(as.data.table(p))
}
\concept{Prediction}
