% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorSimsurv.R
\name{TaskGeneratorSimsurv}
\alias{TaskGeneratorSimsurv}
\alias{mlr_task_generators_simsurv}
\title{Survival Task Generator for Package 'simsurv'}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:TaskGenerator]{mlr3::TaskGenerator}.}
\description{
A \link[mlr3:TaskGenerator]{mlr3::TaskGenerator} calling \code{\link[simsurv:simsurv]{simsurv::simsurv()}} from package \CRANpkg{simsurv}.

This generator currently only exposes a small subset of the flexibility of \CRANpkg{simsurv}, and just creates a small data set with the following numerical covariates:
\itemize{
\item \code{treatment}: Bernoulli distributed with log hazard ratio \code{-0.5}.
\item \code{height}: Normally distributed with log hazard ratio \code{1}.
\item \code{weight}: normally distributed with log hazard ratio \code{0}.
}

See \code{\link[simsurv:simsurv]{simsurv::simsurv()}} for an explanation of the hyperparameters.
}
\section{Construction}{
\preformatted{TaskGeneratorSimsurv$new()
mlr_task_generators$get("simsurv")
tgen("simsurv")
}
}

\examples{
generator = mlr3::mlr_task_generators$get("simsurv")
task = generator$generate(20)
task$head()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:TaskGenerator]{TaskGenerators}: \link[mlr3:mlr_task_generators]{mlr3::mlr_task_generators}
}
