% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSurv.R
\docType{data}
\name{TaskSurv}
\alias{TaskSurv}
\title{Survival Task}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Task}/\link{TaskSupervised}.}
\description{
This task specializes \link[mlr3:Task]{mlr3::Task} and \link[mlr3:TaskSupervised]{mlr3::TaskSupervised} for right-censored survival problems.
The target column is assumed to be a factor.
Predefined tasks are stored in \link[mlr3:mlr_tasks]{mlr3::mlr_tasks}.

The \code{task_type} is set to \code{"surv"}.
}
\section{Construction}{
\preformatted{t = TaskSurv$new(id, backend, time, status)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Name of the task.
\item \code{backend} :: \link{DataBackend}
\item \code{time} :: \code{numeric()}\cr
Event times.
\item \code{status} :: \code{integer()} | \code{logical()}\cr
Event indicator. "0"/\code{FALSE} means alive (no event), "1"/\code{TRUE} means dead (event).
}
}

\section{Fields}{

See \link[mlr3:TaskSupervised]{mlr3::TaskSupervised}.
}

\section{Methods}{

All methods from \link[mlr3:TaskSupervised]{mlr3::TaskSupervised}, and additionally:
\itemize{
\item \code{survfit(strata = character())}\cr
\code{character()} -> \code{\link[survival:survfit]{survival::survfit()}}\cr
Creates a \code{\link[survival:survfit]{survival::survfit()}} object for the survival times.
Argument \code{strata} can be used to stratify into multiple groups.
}
}

\examples{
library(mlr3)
lung = mlr3misc::load_dataset("lung", package = "survival")
lung$status = (lung$status == 2L)
b = as_data_backend(lung)
task = TaskSurv$new("lung", backend = b, time = "time", status = "status")

task$target_names
task$feature_names
task$formula()
task$truth()
task$survfit("age > 50")
}
\concept{Task}
\keyword{datasets}
