% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_scores.R
\name{weighted_survival_score}
\alias{weighted_survival_score}
\title{\itemize{
\item \code{tmax_apply} = TRUE => one of \code{t_max}, \code{p_max} is given
\item \code{tmax_apply} = FALSE => \code{times} is given
The \code{t_max} cutoff will be applied later in the predicted survival matrix
to filter observations (rows) and time points (columns) + filter the
(time, status) target on both train (if provided) and test data
}}
\usage{
weighted_survival_score(
  loss,
  truth,
  distribution,
  times = NULL,
  t_max = NULL,
  p_max = NULL,
  proper,
  train = NULL,
  eps,
  ...
)
}
\description{
\itemize{
\item \code{tmax_apply} = TRUE => one of \code{t_max}, \code{p_max} is given
\item \code{tmax_apply} = FALSE => \code{times} is given
The \code{t_max} cutoff will be applied later in the predicted survival matrix
to filter observations (rows) and time points (columns) + filter the
(time, status) target on both train (if provided) and test data
}
}
